% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{acquireDirectoryStructure}
\alias{acquireDirectoryStructure}
\title{Acquire structure of input directory}
\usage{
acquireDirectoryStructure(dataFolder, coordinateFilenames, tasFilenames,
  timeFilenames, models_to_run, dataDirectories, threshold_ensemble)
}
\arguments{
\item{dataFolder}{Character string with pathway to directory that
contains climate projection data.}

\item{coordinateFilenames}{Character string with filename of the file
containing the latitude and longitude coordinates in the bottom
directory of each ensemble member subdirectory. (See the package
vignette for an example of the required structure for this file.)}

\item{tasFilenames}{Character string with filename of the file
containing the climate projection data in the bottom
directory of each ensemble member subdirectory. (See the package
vignette for an example of the required structure for this file.)}

\item{timeFilenames}{Character string with filename of the file
containing the date dataframe in the bottom
directory of each ensemble member subdirectory. (See the package
vignette for an example of the required structure for this file.)}

\item{models_to_run}{A character vector with either "all" (the default),
in which case the function runs through all models in the specified data
directory, or the names of the models to run, using the names of each
model's subdirectory within the data directory (e.g.,
\code{c("bcc1", "ccsm")})}

\item{dataDirectories}{A list object, with two elements for each of the
two subdirectories included in the main directory. Typically, these will
be a historical directory and a climate projection directory. Each element
of the list should be named with the names of the subdirectories and
should provide a numeric vector with the starting year and ending year of
the data within each of the two subdirectories (e.g.,
\code{list("historical" = c(1980, 2004), "rcp85" = c(2006, 2099))})}

\item{threshold_ensemble}{A character vector giving the name of the ensemble
member that should be used when determining the threshold temperature for
the heatwave definition for each community for each climate model (e.g.,
\code{"r1i1p1"})}
}
\value{
A list object outlining the file structure of the directory
   containing the climate projections.This list has an element
   for each climate model (e.g.,"bcc1"). The first element within each
   of these elements is the name of the model. The second element within
   the first-level element gives the file paths for location grids,
   climate projections, and projection times for each ensemble run of the
   model.
}
\description{
This function walks through the directory structure of the user-provided
directory of climate projection files and parses out a list of the
climate models and ensemble members included in that directory based
on subdirectory names.
}
\note{
The files in the bottom directory must all have a .csv extension
   and must have names corresponding to the relevant "*Filenames" parameter
   of this function. All other files will be removed when creating the
   directory structure.
}

