% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyHeatwaves.R
\name{apply_all_models}
\alias{apply_all_models}
\title{Apply a function across all projections}
\usage{
apply_all_models(out, FUN, city_specific = FALSE)
}
\arguments{
\item{FUN}{A character string giving the name of a function to apply to
the heatwave dataframe located in the file specified by \code{hwPath}.
This function must only take one argument, \code{hw_datafr}, which
identifies a dataframe as generated by \code{\link{gen_hw_set}},
with the columns for staring and ending dates in the Date class,
the column for the city as a factor, and all other columns as
numeric class. The function should output a single value when
applied to the full dataframe.}

\item{city_specific}{TRUE / FALSE: Should the function be calculated
separately for each community within the dataframe?}
}
\value{
A dataframe with the value output by the user-specified
   function, as applied to every dataframe of heatwaves and characteristics
   for every ensemble member of every climate model.
}
\description{
This function will take a user-specified function and apply it across
all the heatwave projection data files output by \code{\link{gen_hw_set}}.
It will generate either a single value for every ensemble member within
every climate model, if \code{city_specific} is set to \code{FALSE}, or
a value for every community for every ensemble member, if
\code{city_specific} is set to \code{TRUE}.
}

