% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyHeatwaves.R
\name{apply_hw_projections}
\alias{apply_hw_projections}
\title{Apply a function to projected heatwaves}
\usage{
apply_hw_projections(hwPath, FUN, city_specific = FALSE)
}
\arguments{
\item{hwPath}{A filepath to a comma-separated (csv) file with a dataset
of heatwaves and their characteristics, as generated by
\code{\link{gen_hw_set}}. The file at the specified filepath must
exactly conform to the format of the heatwave files created by
\code{\link{gen_hw_set}}.}

\item{FUN}{A character string giving the name of a function to apply to
the heatwave dataframe located in the file specified by \code{hwPath}.
This function must only take one argument, \code{hw_datafr}, which
identifies a dataframe as generated by \code{\link{gen_hw_set}},
with the columns for staring and ending dates in the Date class,
the column for the city as a factor, and all other columns as
numeric class. The function should output a single value when
applied to the full dataframe.}

\item{city_specific}{TRUE / FALSE: Should the function be calculated
separately for each community within the dataframe?}
}
\description{
This function takes a user-specified function and applies is to a
single file of heatwave projections, as specified by \code{hwPath}.
}
\note{
The function input as \code{FUN} must follow a very specific
   structure. It most have only one argument, and that argument
   must be a dataframe with heatwaves and their characteristics, as
   generated by the \code{\link{gen_hw_set}} function.
}

