% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleApplyFunctions.R
\name{average_length}
\alias{average_length}
\title{Calculate average length of heatwaves}
\usage{
average_length(hw_datafr)
}
\arguments{
\item{hw_datafr}{A dataframe of heatwaves and their characteristics, as
created by \code{\link{gen_hw_set}}}
}
\value{
A numeric value giving the average length of heatwaves in the
   dataframe
}
\description{
This function takes a dataframe of heatwaves, as created by
\code{\link{gen_hw_set}}, and calculates the average length (in days) of
heatwaves in the dataframe.
}

