% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleApplyFunctions.R
\name{average_mean_temp}
\alias{average_mean_temp}
\title{Calculate average temperature of heatwaves}
\usage{
average_mean_temp(hw_datafr)
}
\arguments{
\item{hw_datafr}{A dataframe of heatwaves and their characteristics, as
created by \code{\link{gen_hw_set}}}
}
\value{
A numeric value giving the average of the mean daily
   temperature during each heatwave across all heatwaves in the dataframe,
   in degrees Fahrenheit
}
\description{
This function takes a dataframe of heatwaves, as created by
\code{\link{gen_hw_set}}, and calculates the average of the mean daily
temperature during each heatwave across all heatwaves in the dataframe.
}

