% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{buildStructureExperiments}
\alias{buildStructureExperiments}
\title{Generate file structure for an experiment}
\usage{
buildStructureExperiments(modelName, experiment, dataPath, coordinateFilenames,
  tasFilenames, timeFilenames, dataDirectories)
}
\arguments{
\item{modelName}{Character string of climate model name (e.g., "bcc1"). This
name is typically generated for use in this function from the subdirectory
names for the climate model within the directory of projection data
specified by the user in \code{\link{gen_hw_set}}.}

\item{experiment}{Character string of the experiment of interest.
Possible variables are "historical" or "rcp85".}

\item{dataPath}{Character string of the file path to the directory
containing the climate projections. Must include the final `/`.}

\item{coordinateFilenames}{Character string with filename of the file
containing the latitude and longitude coordinates in the bottom
directory of each ensemble member subdirectory. (See the package
vignette for an example of the required structure for this file.)}

\item{tasFilenames}{Character string with filename of the file
containing the climate projection data in the bottom
directory of each ensemble member subdirectory. (See the package
vignette for an example of the required structure for this file.)}

\item{timeFilenames}{Character string with filename of the file
containing the date dataframe in the bottom
directory of each ensemble member subdirectory. (See the package
vignette for an example of the required structure for this file.)}

\item{dataDirectories}{A list object, with two elements for each of the
two subdirectories included in the main directory. Typically, these will
be a historical directory and a climate projection directory. Each element
of the list should be named with the names of the subdirectories and
should provide a numeric vector with the starting year and ending year of
the data within each of the two subdirectories (e.g.,
\code{list("historical" = c(1980, 2004), "rcp85" = c(2006, 2099))})}
}
\value{
A list that is the length of the number of ensembles. Each element
   is a return value of the \code{\link{buildStructureEnsembles}} function for
   one of the ensemble members in that experiment of that climate model
   within the user-specified projections directory.
}
\description{
This function generates a list object with the file structure of files
within the user-specified projection directory for a single experiment
(i.e., "historical" or "rcp85").
}

