% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{buildStructureModels}
\alias{buildStructureModels}
\title{Generate list of file structure}
\usage{
buildStructureModels(model, experiments, all, dataFolder, coordinateFilenames,
  tasFilenames, timeFilenames)
}
\arguments{
\item{model}{Character string of climate model name}

\item{experiments}{Character vector of the experiment(s) of interest.
Possible variables are "historical", "rcp85", or both.}

\item{all}{Character vectors with the relative pathnames of all
climate projection files of interest from the directory
specified in dataPath}

\item{dataPath}{Character string of the file path to the directory
containing the climate projections. Must include the final `/`.}
}
\value{
A list of length 3. The first element is the name of the model
   whose structure was being built. The second element is the historical
   experiment hierarchy. The third element is the hierarchy of the future
   projection directory for the model. The second and third elements are
   return values of \code{buildStructureExperiments}.
}
\description{
This function takes input from \code{acquireDirectoryStructure} and
   uses it to generate a list object with the projection directory
   file structure. This parsed file structure is later used to lead other
   code through all climate models and ensemble members in the input
   projection directory.
}
\examples{
model <- "bcc1"
experiments <- c("historical", "rcp85")
all <- c("rcp85/bcc1/r1i1p1/latitude_longitude_NorthAmerica_12mo.csv",
         "rcp85/bcc1/r1i1p1/tas_NorthAmerica_12mo.csv",
         "rcp85/bcc1/r1i1p1/time_NorthAmerica_12mo.csv")

dataFolder <- system.file("cmip5", package = "futureheatwaves")
dataFolder <- paste0(dataFolder, "/")

coordinateFilenames <- "latitude_longitude_NorthAmerica_12mo.csv"
tasFilenames <- "tas_NorthAmerica_12mo.csv"
timeFilenames <- "time_NorthAmerica_12mo.csv"

buildStructureModels(model = model,
                     experiments = experiments,
                     all = all,
                     dataFolder,
                     coordinateFilenames = coordinateFilenames,
                     tasFilenames = tasFilenames,
                     timeFilenames = timeFilenames)
}

