% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{checkCustomBounds}
\alias{checkCustomBounds}
\title{Check year boundaries for errors}
\usage{
checkCustomBounds(boundList, dataDirectories)
}
\arguments{
\item{boundList}{Either FALSE, if user wishes to use the default
boundaries, or a set of boundary years in the format
c(historical period earliest year, historical period latest year,
future projection period earliest year, future projection period
latest year).}

\item{dataDirectories}{A list object, with two elements for each of the
two subdirectories included in the main directory. Typically, these will
be a historical directory and a climate projection directory. Each element
of the list should be named with the names of the subdirectories and
should provide a numeric vector with the starting year and ending year of
the data within each of the two subdirectories (e.g.,
\code{list("historical" = c(1980, 2004), "rcp85" = c(2006, 2099))})}
}
\description{
This function inputs the boundary lists specified in
\code{\link{gen_hw_set}}, \code{dataBoundaries} and
\code{referenceBoundaries}, and checks them for errors in structure of the
input or in the years selected.
}

