% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCities.R
\name{checkLocations}
\alias{checkLocations}
\title{[Short title for function]}
\usage{
checkLocations(locationsList)
}
\arguments{
\item{locationList}{[Whis is this?]}
}
\value{
[What does this return?]
}
\description{
[Short title for function]
}

