% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{closest_point}
\alias{closest_point}
\title{Find closest grid point to a city location}
\usage{
closest_point(city, latlong)
}
\arguments{
\item{city}{A numeric vector containing the city's ID, latitude, and
longitude, in that order.}

\item{latlong}{A dataframe of latitude and longitude pairings for each
climate grid cell location, with latitudes in the first column and
longitudes in the second column.}
}
\value{
An index corresponding to the column in the climate projections
   dataframe for that climate model that is closest to the city
   coordinates.
}
\description{
This function identifies the closest grid points in a climate model to a
   community based on the community's latitude and longitude using the
   Pythagorean theorem.
}

