% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHeatwaves.R
\name{consolidate}
\alias{consolidate}
\title{Stitch all heatwave lists together}
\usage{
consolidate(hwDataframeList)
}
\arguments{
\item{hwDataframeList}{[What is this?]}
}
\value{
Dataframe with all heatwaves for the cities of interest
   for a climate projection.
}
\description{
Stitch all heatwave lists together
}

