% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create.heatwave.dataset}
\alias{create.heatwave.dataset}
\title{Create heatwave dataset for a directory of climate projections}
\usage{
create.heatwave.dataset(out, dataFolder, citycsv, RorCPP = 1,
  IDheatwavesReplacement = FALSE, dataBoundaries = FALSE,
  referenceBoundaries = FALSE)
}
\arguments{
\item{out}{Character string with pathway to directory to which
heatwave files will be written.}

\item{dataFolder}{Character string with pathway to directory that
contains climate projections. Must include the final backslash.}

\item{citycsv}{Character string giving the filepath to a .csv
file with latitude and longitude values for each city.}

\item{RorCPP}{0 /1 flag that indicates whether to use R (1) or
CPP (0) functions}

\item{IDheatwavesReplacement}{Either FALSE, to use the default
heatwave definition, or a user-specified custom function to
use to identify heatwaves.}

\item{dataBoundaries}{Custom time boundaries for extracting data
from the ensembles. Format: c(historical low bound, historical
high bound, reference low bound, reference high bound).
Restrictions: Bounds cannot span multiple experiments}

\item{referenceBoundaries}{Reference boundaries.}
}
\value{
[What does this function return?]
}
\description{
Create heatwave dataset for a directory of climate projections
}

