% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{createAccumulators}
\alias{createAccumulators}
\title{Create accumulators}
\usage{
createAccumulators()
}
\value{
A closure that maintains data structures that track
information about the models and the location vectors for each
ensemble respectively.
}
\description{
Create accumulators
}

