% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datafr}
\alias{datafr}
\title{Example data to input to heat wave identifier functions}
\format{A dataframe with 7,300 rows and 2 columns. The variables are:
\itemize{
  \item date: A Date vector with the date of the projection
  \item tmpd: A numeric vector with the temperature of the projection, in
              degrees Fahrenheit
}}
\usage{
datafr
}
\description{
This dataset provides an example of the dataframe to be input into a
custom heat wave identification function. This dataset can be used to test a
custom heat wave identification function before using the function
in \code{\link{gen_hw_set}}.
}
\keyword{datasets}

