% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{formDates}
\alias{formDates}
\title{Determine dates for the requested time period.}
\usage{
formDates(times, bounds)
}
\arguments{
\item{times}{A dataframe containing the time data for one of the
climate model ensemble's projection.}

\item{bounds}{A numeric vector with the desired time boundaries of the
ensemble.}
}
\value{
A vector of dates formatted as year-month-day (%Y-%m-%d) spanning
   the desired time boundaries.
}
\description{
Acquire the dates spanning the desired time period of the ensemble
}

