% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futureheatwaves.R
\docType{package}
\name{futureheatwaves}
\alias{futureheatwaves}
\alias{futureheatwaves-package}
\title{futureheatwaves: Identify and Characterize Heatwaves in Climate Projections}
\description{
futureheatwaves Takes locally saved climate projections and identifies and
characterizes heatwaves in those projections. Heatwaves are defined
based on a definition frequently used in climate epidemiology.
Characterizations include several metrics of heatwave length,
intensity, and timing in the summer.
}

