% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gen_hw_set}
\alias{gen_hw_set}
\title{Create and write heatwave projections}
\usage{
gen_hw_set(out, dataFolder, citycsv, coordinateFilenames, tasFilenames,
  timeFilenames, IDheatwavesReplacement = "IDHeatwavesR",
  thresholdBoundaries = c(1981, 2004), projectionBoundaries = c(2061, 2080),
  referenceBoundaries = c(2061, 2080), probThreshold = 0.98,
  printWarning = TRUE)
}
\arguments{
\item{out}{Character string with pathway to directory to which
heatwave files will be written.}

\item{dataFolder}{Character string with pathway to directory that
contains climate projection data.}

\item{citycsv}{Character string giving the filepath to a .csv
file with latitude and longitude values for each city.}

\item{coordinateFilenames}{Character string with name of the files
containing the latitude and longitude coordinates
corresponding to the columns of the time series data.}

\item{tasFilenames}{Character sting with name of files containing the time
series data.}

\item{timeFilenames}{Character string with name of the files containing the
date information corresponding to the rows of the time series data.}

\item{IDheatwavesReplacement}{Either FALSE, to use the default
heatwave definition, or a user-specified custom function to
use to identify heatwaves.}

\item{thresholdBoundaries}{Custom time boundaries for determining
threshold temperatures for the heatwave definition.
Format: c(earliest year, latest year).
Restrictions: Bounds must be between the years 1980 and 2099 and
cannot span 2005 (i.e., both must either be before 2005 or after 2005).}

\item{referenceBoundaries}{Custom time boundaries to use in calculating
relative characteristics for heatwaves (i.e., to use when exploring
the role of adaptation in projections).
Format: c(earliest year, latest year).
Restrictions: Bounds must be between the years 1980 and 2099 and
cannot span 2005 (i.e., both must either be before 2005 or after 2005).}

\item{probThreshold}{Numerical value between 0 and 1 specifying the threshold
of temperature to use when defining heatwaves. The default value is 0.98
(i.e., a heatwave is two or more days above the community's 98th
percentile temperature).}

\item{printWarning}{TRUE / FALSE specifying whether to print out the progress
of the function as it runs. Default is TRUE.}

\item{thresholdBoundaries}{Custom time boundaries for projections.
Format: c(earliest year, latest year).
Restrictions: Bounds must be between the years 1980 and 2099 and
cannot span 2005 (i.e., both must either be before 2005 or after 2005).}
}
\value{
This function returns a dataframe listing the name of each climate
   model included in the directory of projection files inputted to the
   function as well as the number of historical and future projection
   ensembles included for each model. This output can be used as a check
   that the function processed through the directory of input files
   specified using the \code{dataFolder} argument.
}
\description{
This function creates heatwave projection datasets for all models and
ensemble members in a directory of climate projections for a specified
set of communities.The resulting heatwave projections are written out
to a specified directory on the user's local computer.
}

