% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleApplyFunctions.R
\name{heatwave_days}
\alias{heatwave_days}
\title{Calculate total heatwave days}
\usage{
heatwave_days(hw_datafr)
}
\arguments{
\item{hw_datafr}{A dataframe of heatwaves and their characteristics, as
created by \code{\link{gen_hw_set}}}
}
\value{
A numeric value giving the number of heatwave days in the dataframe
}
\description{
This function takes a dataframe of heatwaves, as created by
\code{\link{gen_hw_set}}, and calculates the total number of heatwave days in
the dataframe (sum of the number of days in each heatwave for all heatwaves
in the dataset).
}

