% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hw_datafr}
\alias{hw_datafr}
\title{Example of heatwave characteristics dataset}
\format{A dataframe with 45 rows and 23 columns. The variables are:
\itemize{
   \item hw.number: A sequential number identifying each heatwave in each
       community
   \item mean.temp: Average of mean daily temperature across all days in the
      heatwave
   \item max.temp: Highest value of mean daily temperature across days in the
      heatwave
   \item min.temp: Lowest value of mean daily temperature across days in the
      heatwave
   \item length: Number of days in the heatwave
   \item start.date: Date of the first day of the heatwave
   \item end.date: Date of the last day of the heatwave
   \item start.doy: Numeric day of the year of the first day of the heatwave
      (1 = Jan. 1, etc.)
   \item start.month: Numeric value indicating the month in which the
       heatwave started (1 = January);
   \item days.above.80: Number of days in the heatwave above 80 degrees
       Fahrenheit
   \item days.above.85: Number of days in the heatwave above 85 degrees
       Fahrenheit
   \item days.above.90: Number of days in the heatwave above 90 degrees
       Fahrenheit
   \item days.above.95: Number of days in the heatwave above 90 degrees
       Fahrenheit
   \item days.above.99th:Number of days in the heatwave above the 99th
       percentile temperature for the community, using the period specified
       by the user with the \code{referenceBoundaries} argument in
       \code{\link{gen_hw_set}} as a reference for determining these
       percentiles
   \item days.above.99.5th: Number of days in the heatwave above the 99.5th
       percentile temperature for the community
   \item first.in.season: Whether the heatwave was the first to occur in its
       season (as determined by whether the heatwave was the first in its
       calendar year)
   \item threshold.temp: The temperature used as the threshold for the
       heatwave definition in the selected community
   \item mean.temp.quantile: The percentile of the average daily mean
       temperature during the heatwave compared to the community's year-round
       temperature distribution, based on the temperatures for the community
       during the period specified by the \code{referenceBoundaries}
       argument in \code{\link{gen_hw_set}}
   \item max.temp.quantile: The percentile of the highest daily mean
       temperature during the heatwave compared to the community's year-round
       temperature distribution
   \item min.temp.quantile: The percentile of the lowest daily mean
       temperature during the heatwave compared to the community's year-round
       temperature distribution
   \item mean.temp.1: The community's average year-round temperature, based
       on the temperatures for the community during the period specified by
       the \code{referenceBoundaries} argument in
       \code{\link{gen_hw_set}}
   \item mean.summer.temp: The community's average May--September
       temperature, based on the temperatures for the community during the
       period specified by the \code{referenceBoundaries} argument
       in \code{\link{gen_hw_set}}
   \item city: The identifier for the community, as given in the file
       specified in the \code{citycsv} argument of
       \code{\link{gen_hw_set}}
}}
\usage{
hw_datafr
}
\description{
This dataset provides an example of the structure of the dataframe of
heatwaves and their characteristics, as created by \code{\link{gen_hw_set}}.
This example dataset can be used in developing custom functions to use with
\code{\link{apply_all_models}}.
}
\keyword{datasets}

