% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDHeatwaves.R
\name{IDHeatwavesCPPwrapper}
\alias{IDHeatwavesCPPwrapper}
\title{Identify heatwaves in a time series}
\usage{
IDHeatwavesCPPwrapper(datafr, threshold)
}
\arguments{
\item{datafr}{A dataframe with daily temperature projections in the
the community being processed. This dataframe must have two columns:
(1) the first column must have the date of each observation, with
class "Date" and; (2) the second column must have temperatures
(in typical runs of this function, this will be in degrees
Fahrenheit). In the normal running of this package, this dataframe
will be generate by the closure created by
\code{\link{createCityProcessor}}.}

\item{threshold}{Numeric string with threshold temperature used in
the heatwave definition, in the same units as the temperature time
series in \code{datafr}.}
}
\value{
Returns the dataframe entered as \code{datafr}, but with new
   columns providing heatwave identifiers. The returned dataframe will
   have new columns for whether a day was part of a heatwave (\code{hw},
   0 / 1), if it was part of a heatwave, the number of the heatwave
   (\code{hw.number}), and whether the day was the first day in a heatwave
   (\code{first.hw.day}, 0 /1).
}
\description{
This function identifies heatwaves in a time series of temperature
data using a heatwave definition that a heatwave must be two or more
days with temperatures equal to or above some threshold temperature.
}
\note{
This function gives identical results to the \link{\code{IDHeatwavesR}}
   function, but should take less time to run.
}

