% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDHeatwavesForCountry.R
\name{IDHeatwavesR}
\alias{IDHeatwavesR}
\title{[Short title for function]}
\usage{
IDHeatwavesR(city = stop("Error: unspecified city"),
  threshold = stop("Error: unspecified threshold"), days = 2,
  datafr = stop("Error: 'datafr' unspecified"))
}
\arguments{
\item{city}{Character string with city ID.}

\item{threshold}{Numeric string with threshold percentile used in
heatwave definition.}

\item{days}{Numeric string specifying minimum number of days required
in heatwave definition.}

\item{datafr}{Dataframe with daily temperatures in city of interest.}
}
\value{
Returns the dataframe entered as \code{datafr}, but with new
   columns providing heatwave idendifiers.
}
\description{
[Short title for function]
}

