% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{listCustom}
\alias{listCustom}
\title{Create list of custom settings.}
\usage{
listCustom(IDheatwavesReplacement = FALSE, dataBoundaries = FALSE,
  referenceBoundaries = FALSE)
}
\arguments{
\item{IDheatwavesReplacement}{Either FALSE, to use the default
heatwave definition, or a user-specified custom function to
use to identify heatwaves.}

\item{dataBoundaries}{Custom time boundaries for extracting data
from the ensembles. Format: c(historical low bound, historical
high bound, reference low bound, reference high bound).
Restrictions: Bounds cannot span multiple experiments}

\item{referenceBoundaries}{Reference boundaries.}
}
\value{
A list with slots for custom settings specified by the user.
}
\description{
Create list of custom settings.
}

