% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{listGlobal}
\alias{listGlobal}
\title{Creates lists for 'global' variables and custom data respectively.}
\usage{
listGlobal(out = FALSE, dataFolder = FALSE, citycsv = FALSE,
  RorCPP = FALSE)
}
\arguments{
\item{out}{Character string with pathway to directory to which
heatwave files will be written.}

\item{dataFolder}{Character string with pathway to directory that
contains climate projections. Must include the final backslash.}

\item{citycsv}{Character string giving the filepath to a .csv
file with latitude and longitude values for each city.}

\item{RorCPP}{0 /1 flag that indicates whether to use R (1) or
CPP (0) functions.}
}
\value{
A list with slots for the output directory pathname,
   the pathname for the directory with the climate projections,
   the pathname with the csv with city latitudes and longitudes,
   and the RorCPP choices.
}
\description{
Creates lists for 'global' variables and custom data respectively.
}
\note{
These functions exist to give the user a way to create the
   required "global" and "custom" parameters required by the rest
   of the program without running the entire program.
}
\examples{
out <- "~/Downloads/sample/results"
dataFolder <- "~/Downloads/sample/cmip5/"
citycsv <- "inst/cities.csv"
RorCPP <- FALSE
listGlobal(out, dataFolder, citycsv, RorCPP)
}

