% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCities.R
\name{makeMap}
\alias{makeMap}
\title{Map cities of interest}
\usage{
makeMap(locationsList, cities, sampleEnsemble)
}
\arguments{
\item{sampleEnsemble}{}
}
\value{
Checks locations in locationList for ... and, if ...,
   plots a map of ... .
}
\description{
Map cities of interest
}

