% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCities.R
\name{map_grid_ggmap}
\alias{map_grid_ggmap}
\title{Create a map of model grid using ggmap}
\usage{
map_grid_ggmap(plot_model, out, lon_transform = TRUE)
}
\arguments{
\item{plot_model}{A character string with the name of the model to plot}

\item{out}{Character string with pathname to which heat wave files were
written by \code{\link{gen_hw_set}}. Typically, this will be the same
pathname as that specified with \code{out} when running
\code{\link{gen_hw_set}}.}

\item{lon_transform}{Logical value indicating whether longitude is in a
format that needs to be transformed. Often, U.S. locations will give
longitude as an absolute value, which needs to be transformed using
\eqn{360 - longitude} before mapping. If using longitude values that
do not require transformation, set this value to FALSE.}
}
\value{
A \code{ggplot2} object with a map of grid points for the climate
   model that were used in processing heat waves for the study locations,
   with a line drawn from each study locations to the grid point used for
   it.
}
\description{
This function creates a map of the grid points of a climate model used for
the study locations and draws lines connecting each study city to its climate
model grid point, using as a base a map pulled from Google Maps with
\code{ggmap}.
}
\note{
This function creates a \code{ggplot2} object, so the output can be
   edited using \code{ggplot2} functions.
}
\examples{
out <- system.file("extdata/example_results", package = "futureheatwaves")
map_grid_ggmap(plot_model = "bcc1", out = out)

}

