% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCities.R
\name{mapCities}
\alias{mapCities}
\title{Map locations and cities}
\usage{
mapCities(locations, cities, sampleEnsemble)
}
\arguments{
\item{locations}{[What is this?]}

\item{cities}{Dataframe with a row for every city and columns for
city ID, latitude, and longitude.}

\item{sampleEnsemble}{[What is this?]}
}
\value{
Plots  a map of all cities in the \code{cities} dataframe,
    as well as ...
}
\description{
Map locations and cities
}

