% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleApplyFunctions.R
\name{number_of_heatwaves}
\alias{number_of_heatwaves}
\title{Calculate number of heatwaves}
\usage{
number_of_heatwaves(hw_datafr)
}
\arguments{
\item{hw_datafr}{A dataframe of heatwaves and their characteristics, as
created by \code{\link{gen_hw_set}}.}
}
\value{
A numeric value giving the number of heatwaves in the dataframe.
}
\description{
This function takes a dataframe of heatwaves, as created by
\code{\link{gen_hw_set}}, and calculates the number of heatwaves in
the dataframe.
}
\details{
To calculate the number of heatwaves, this function determines the number
of rows in the dataframe, since \code{\link{gen_hw_set}} outputs a
dataframe with one and only one row per heatwave.
}

