% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{process_cities_file}
\alias{process_cities_file}
\title{Process city file once it's read}
\usage{
process_cities_file(cities, lat_lon_colnames)
}
\arguments{
\item{cities}{Dataframe with communities you want to analyze and their latitudes
and longitudes.The dataframe must have a column named \code{city} with a unique
identifier for each community in the study, as well as columns for latitude and
longitude.Other columns may be included in the dataset, but will not be run
within this code.}

\item{lat_lon_colnames}{A character vector of length two with the column names
in the \code{cities} dataframe for latitude (first vector element) and
longitude (second vector element)}
}
\description{
This function takes the dataframe read in fram the user-specified file. It then
renames the columns for latitude and longitude \code{lat} and \code{lon} based on
the user's selections in the "lat_lon_colnames" option for \code{gen_hw_set}. If
there are extra columns besides those and the \code{city} column, this function
removes them.
}

