% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processEnsemble}
\alias{processEnsemble}
\title{Extract the desired data out of the ensemble}
\usage{
processEnsemble(ensemble, experiment, modelName, global, custom, reference)
}
\arguments{
\item{ensemble}{Character string that gives the absolute file path
for the directory with a particular ensemble member of a climate
model projection.}

\item{experiment}{Character string of the experiment of interest.
Possible variables are "historical" or "rcp85".}

\item{global}{List of global parameters generated by listGlobal.}

\item{custom}{List of custom parameters generated by listCustom.}

\item{reference}{[What is this parameter?]}
}
\value{
Returns a list with...
}
\description{
Extract the desired climate data out of a single ensemble member of
the climate model's projections.
}

