% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processHistorical}
\alias{processHistorical}
\title{Calculate threshold temperatures from historical projections}
\usage{
processHistorical(model, global, custom, reference = FALSE)
}
\arguments{
\item{model}{Character string of the climate model (e.g., "bcc1").}

\item{global}{List of global parameters generated by listGlobal.}

\item{custom}{List of custom parameters generated by listCustom.}

\item{reference}{[What is this parameter?]}
}
\value{
Writes threshold temperatures to file and returns ... ?
}
\description{
Calculate the threshold temperatures for the model using the
model's r1i1p1 historical ensemble and saves these to file.
}

