% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processProjections}
\alias{processProjections}
\title{Create heatwave dataframe for climate projection}
\usage{
processProjections(ensemble, modelName, ensembleWriter, thresholds, global,
  custom, accumulators, reference = reference,
  reference_dates = reference_dates)
}
\arguments{
\item{ensemble}{List with parsed information about the directory structure for
a specific ensemble member from the user-specified projections directory.
This list is a subset of the list generated by
\code{\link{acquireDirectoryStructure}}.}

\item{modelName}{Character string of climate model name (e.g., "bcc1"). This
name is typically generated for use in this function from the subdirectory
names for the climate model within the directory of projection data
specified by the user in \code{\link{gen_hw_set}}.}

\item{ensembleWriter}{A closure created by \code{createEnsembleWriter} to
write output from this process to a file within the output directory
specified by the user.}

\item{thresholds}{A vector with the thresholds to use within each community
in that community's heatwave definition.These are typically automatically
determine during the run of the \code{gen_hw_set} function based on
historical climate projections within the community.}

\item{global}{An object name for the 'global' variable list. This
parameter is passed through from \code{\link{gen_hw_set}} and includes user
specifications for the path to the output directory, the path to the
input climate projections, the dataframe with community locations, and the
format for the filenames of the coordinate, tas, and time files in the
ensemble-specific subdirectories of the projection directory input by
the user.}

\item{custom}{An object name for the 'custom` data list. This parameter is
passed through from \code{\link{gen_hw_set}} and includes user specifications
for, if specified as other than the default, the name of the R function to
use to identify heatwaves, alternative upper and lower year boundaries
for the data used to determine threshold temperatures for the heatwave
definition, alternative upper and lower year boundaries
for the projection period of the heatwave datasets being generated,
alternative upper and lower year boundaries for the reference period
to be used when measuring all heatwave characteristics related to
relative, rather than absolute, temperature, and the percentile
threshold to use when defining heatwaves.}

\item{accumulators}{An object name for the closure generated by
\code{\link{createAccumulators}}.}

\item{reference}{FALSE, if the user has not specified custom reference
boundaries through the \code{referenceBoundaries} argument in
\code{\link{gen_hw_set}}, otherwise a vector with four elements
(earliest historical year, latest historical year,
earliest future projection year, latest future projection year)
giving the user-specified custom date boundaries for the
heatwave projections.}
}
\value{
Write every heatwave dataframe to a .csv and return the ensemble
   used as the reference.
}
\description{
For each reference ensemble, create the heatwave dataframe and
write it to file. Also store the locations  vector for each ensemble.
}

