% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readLatLong}
\alias{readLatLong}
\title{Read the latitude and longitude data for a given ensemble}
\usage{
readLatLong(ensemble)
}
\arguments{
\item{ensemble}{Ensemble component of the directory structure that contains the path
to the latitude and longitude data to be read}
}
\value{
A dataframe of the latitude and longitude data of the ensemble
}
\description{
Read the latitude and longitude data for a given ensemble
}

