% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readtas}
\alias{readtas}
\title{Read climate projection data}
\usage{
readtas(ensemble, global)
}
\arguments{
\item{ensemble}{Character vector that includes
the file paths to (1) the latitude and longitude file; (2) the
climate projection file; and (3) the projection dates file
for the selected climate model.}

\item{global}{An object name for the \code{global} object. This
object is created by \code{\link{gen_hw_set}} and includes user
specifications for the path to the output directory, the path to the
input climate projections, the dataframe with community locations, and the
format for the filenames of the coordinate, tas, and time files in the
ensemble-specific subdirectories of the projection directory input by
the user.}
}
\value{
A dataframe of climate projection data where each column
   corresponds to a climate model grid point and each row corresponds
   to a date of observation.
}
\description{
This function reads climate projection data for a climate
model into R from a locally-stored comma-separated file located within
the directory specified by the \code{dataFolder} argument of
\code{\link{gen_hw_set}}.
}

