% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readTimes}
\alias{readTimes}
\title{Read the date data for a given ensemble}
\usage{
readTimes(ensemble)
}
\arguments{
\item{ensemble}{Ensemble component of the directory structure that contains the path
to the date data to be read}
}
\value{
A dataframe of the date data of the ensemble
}
\description{
Read the date data for a given ensemble
}

