% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{storeHeatwaveEntry}
\alias{storeHeatwaveEntry}
\title{Helper function for IDHeatwavescpp that addes entries to the data structures
that track heatwave information}
\usage{
storeHeatwaveEntry(index, hwSize, hwCounter, hw, hwNumber)
}
\arguments{
\item{index}{An index into the various}

\item{hwSize}{Size of the heat wave to be added.}

\item{hwCounter}{Current number of heat waves.}

\item{hw}{A reference of the vector that contains the heat waves}

\item{hwNumber}{A reference of the vector that contains the heat wave
numbers.}
}
\description{
Helper function for IDHeatwavescpp that addes entries to the data structures
that track heatwave information
}

