% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{storeZeroes}
\alias{storeZeroes}
\title{Helper function that adds zeroes to the data structures that track heatwave information}
\usage{
storeZeroes(index, potentialHeatwave, hw, hwNumber)
}
\arguments{
\item{index}{An index into the various}

\item{potentialHeatwave}{Size of the potential heatwave that turned out not to be a heatwave.}

\item{hw}{A reference of the vector that contains the heatwaves}

\item{hwNumber}{A reference of the vector that contains the heatwave numbers.}
}
\description{
Helper function that adds zeroes to the data structures that track heatwave information
}

