% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{swapPaste}
\alias{swapPaste}
\title{Paste second element before first, separated by /}
\usage{
swapPaste(second, first)
}
\arguments{
\item{second}{Character string with element to paste second in output.}

\item{first}{Character string with element to paste first in output.}
}
\description{
Paste second element before first, separated by /
}
\examples{
second <- "latitude_longitude_NorthAmerica_12mo.csv"
first <- "/Users/brookeanderson/Downloads/sample/cmip5/rcp85/bcc1/r1i1p1"
swapPaste(second, first)
}

