% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{writeAccumulators}
\alias{writeAccumulators}
\title{Write model information to file}
\usage{
writeAccumulators(modelInfoAccumulator, locationList, global)
}
\arguments{
\item{modelInfoAccumulator}{A dataframe with information about the number of
ensembles for each climate model.}

\item{locationList}{A dataframe with information about community locations
and the closest grid point from each climate model to each community.}

\item{global}{An list object created by \code{\link{gen_hw_set}} that
includes user specifications (e.g., the path to the output directory, the
path to the input climate projections, the dataframe with community
locations).}
}
\value{
...
}
\description{
This function writes out a dataframe that accumulates information on all
climate models included in the analysis, including the number of
historical (1980--2004) and future (2006--2099) ensemble members.
}

