% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{writeThresholds}
\alias{writeThresholds}
\title{Write the temperature cutoffs acquired from the r1i1p1 ensemble for each model to a .csv}
\usage{
writeThresholds(modelName, threshOut, thresholds, global, custom)
}
\arguments{
\item{modelName}{Name of the model.}

\item{threshOut}{Name of the .csv file to be written}

\item{thresholds}{Minimum threshold temperatures of each city}

\item{global}{Global list.}

\item{custom}{Custom parameter list.}
}
\description{
Write the temperature cutoffs acquired from the r1i1p1 ensemble for each model to a .csv
}

