% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{acquireDirectoryStructure}
\alias{acquireDirectoryStructure}
\title{Determine climate projection directory structure}
\usage{
acquireDirectoryStructure(dataPath)
}
\arguments{
\item{dataPath}{A character string giving the file path to the
directory with the climate projection data.}
}
\value{
A list object outlining the file structure of the directory
   containing the climate projections.This list has an element
   for each climate model (e.g. ,"bcc1"). The first element within each
   of these elements is the name of the model. The second element with
   the first level element gives the file paths for location grids,
   `tas` and `time` for each ensemble run of the model.
}
\description{
Determine climate projection directory structure
}
\note{
The files in the bottom directory must all have a .csv extension
   and must have one of the following names:
   "latitude_longitude_NorthAmerica_12mo.csv" for the file with grid
   point locations for the climate model,
   "tas_NorthAmerica_12mo.csv" for the file with ...,
   and "time_NorthAmerica_12mo.csv" for the file with ... .
   All other files will be removed when creating the directory structure.
}
\examples{
dataFolder <- "~/Downloads/sample/cmip5/"
finalList <- acquireDirectoryStructure(dataFolder)
str(finalList[[1]][[1]])
str(finalList[[1]][[2]][1])
}

