% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{acquireDirectoryStructure}
\alias{acquireDirectoryStructure}
\title{Acquire structure of input directory}
\usage{
acquireDirectoryStructure(dataFolder, coordinateFilenames, tasFilenames,
  timeFilenames, models_to_run, dataDirectories, threshold_ensemble)
}
\arguments{
\item{dataFolder}{Character string with pathway to directory that
contains climate projection data.}

\item{coordinateFilenames}{Character string with filename of the file
containing the latitude and longitude coordinates in the bottom
directory of each ensemble member subdirectory. (See the package
vignette for an example of the required structure for this file.)}

\item{tasFilenames}{Character string with filename of the file
containing the climate projection data in the bottom
directory of each ensemble member subdirectory. (See the package
vignette for an example of the required structure for this file.)}

\item{timeFilenames}{Character string with filename of the file
containing the date dataframe in the bottom
directory of each ensemble member subdirectory. (See the package
vignette for an example of the required structure for this file.)}
}
\value{
A list object outlining the file structure of the directory
   containing the climate projections.This list has an element
   for each climate model (e.g.,"bcc1"). The first element within each
   of these elements is the name of the model. The second element within
   the first-level element gives the file paths for location grids,
   climate projections, and projection times for each ensemble run of the
   model.
}
\description{
This function walks through the directory structure of the user-provided
directory of climate projection files and parses out a list of the
climate models and ensemble members included in that directory based
on subdirectory names.
}
\note{
The files in the bottom directory must all have a .csv extension
   and must have names corresponding to the relevant "*Filenames" parameter
   of this function. All other files will be removed when creating the
   directory structure.
}

