% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{acquireDirectoryStructure}
\alias{acquireDirectoryStructure}
\title{Acquire structure of input directory}
\usage{
acquireDirectoryStructure(dataFolder, coordinateFilenames, tasFilenames,
  timeFilenames)
}
\arguments{
\item{dataFolder}{A character string giving the file path to the
directory with the climate projection data.}

\item{coordinateFilenames}{Character string with name of the files
containing the latitude and longitude coordinates
corresponding to the columns of the time series data.}

\item{tasFilenames}{Character sting with name of files containing the time
series data.}

\item{timeFilenames}{Character string with name of the files containing the
date information corresponding to the rows of the time series data.}
}
\value{
A list object outlining the file structure of the directory
   containing the climate projections.This list has an element
   for each climate model (e.g.,"bcc1"). The first element within each
   of these elements is the name of the model. The second element within
   the first-level element gives the file paths for location grids,
   climate projections, and projection times for each ensemble run of the
   model.
}
\description{
This function walks through the directory structure of the user-provided
   directory of climate projection files and parses out a list of the
   climate models and ensemble members included in that directory based
   on subdirectory names.
}
\note{
The files in the bottom directory must all have a .csv extension
   and must have names corresponding to the relevant "*Filenames" parameter
   of this function. All other files will be removed when creating the
   directory structure.
}
\examples{
dataFolder <- system.file("cmip5", package = "futureheatwaves")
coordinateFilenames <- "latitude_longitude_NorthAmerica_12mo.csv"
tasFilenames <- "tas_NorthAmerica_12mo.csv"
timeFilenames <- "time_NorthAmerica_12mo.csv"

finalList <- acquireDirectoryStructure(dataFolder = dataFolder,
   coordinateFilenames = coordinateFilenames,
   tasFilenames = tasFilenames,
   timeFilenames = timeFilenames)

str(finalList[[1]][[1]])
str(finalList[[1]][[2]][1])

}

