% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyHeatwaves.R
\name{apply_all_models}
\alias{apply_all_models}
\title{Apply a function across heat waves from all projections}
\usage{
apply_all_models(out, FUN, city_specific = FALSE, ...)
}
\arguments{
\item{out}{Character string with pathname to which heat wave files were
written by \code{\link{gen_hw_set}}.Typically, this will be the same
pathname as that specified with \code{out} when running
\code{\link{gen_hw_set}}.}

\item{FUN}{A character string with the name of a function to apply to
the heat wave dataframes located in the directory specified by
\code{out}. This function must take only the argument, \code{hw_datafr},
a dataframe of identified and characterized heat waves, as generated by
\code{\link{gen_hw_set}}. For an example of one of these dataframes, load
the \code{\link{hw_datafr}} data. The function should output a single
value when applied to the full dataframe.}

\item{city_specific}{TRUE / FALSE: Should the function be applied
separately for each city?}

\item{...}{Optional arguments to FUN}
}
\value{
A dataframe with the value output by the function specified by the
   user with \code{FUN}, as applied to every dataframe of heat waves and
   characteristics for every ensemble member of every climate model.
}
\description{
This function will take a user-specified function and apply it across
all the heat wave files created by \code{\link{gen_hw_set}}.
It will generate either a single value for every ensemble member within
every climate model, if \code{city_specific} is set to \code{FALSE}, or
a value for every city for every ensemble member, if
\code{city_specific} is set to \code{TRUE}.
}
\note{
The function input as \code{FUN} must follow a very specific
   structure. It most have only one argument, and that argument
   must be a dataframe with heat waves and their characteristics, as
   generated by the \code{\link{gen_hw_set}} function. See the
   \code{futureheatwaves} vignette for more guidance on creating and
   applying a custom function to explore the heat waves identified and
   characterized by \code{\link{gen_hw_set}}.
}

