% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleApplyFunctions.R
\name{average_length}
\alias{average_length}
\title{Calculate average length of heat waves}
\usage{
average_length(hw_datafr)
}
\arguments{
\item{hw_datafr}{A dataframe of heat waves and their characteristics, as
created by \code{\link{gen_hw_set}}}
}
\value{
A numeric value with the average length of heat waves in a
   dataframe of heat waves, as generated by \code{\link{gen_hw_set}}.
}
\description{
This function takes a dataframe of heat waves, as created by
\code{\link{gen_hw_set}}, and calculates the average length (in days) of
heat waves in the dataframe.
}
\note{
This function is an example of a function that can be created and
   used to explore heat waves using the \code{\link{apply_all_models}}
   function.
}
\examples{
data(hw_datafr)
average_length(hw_datafr)

}

