% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{buildStructureEnsembles}
\alias{buildStructureEnsembles}
\title{List all files for a single ensemble member}
\usage{
buildStructureEnsembles(ensemble)
}
\arguments{
\item{Ensemble}{character string that gives the absolute file path
for the directory with a particular ensemble member of a climate
model projection.}
}
\description{
List all files for a single ensemble member
}
\examples{
ensemble <- "/Users/brookeanderson/Downloads/sample/cmip5/rcp85/bcc1/r1i1p1"
buildStructureEnsembles(ensemble)
}

