% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{buildStructureEnsembles}
\alias{buildStructureEnsembles}
\title{List all files for a single ensemble member}
\usage{
buildStructureEnsembles(ensemble, coordinateFilenames, tasFilenames,
  timeFilenames)
}
\arguments{
\item{Ensemble}{character string that gives the absolute file path
for the directory with a particular ensemble member of a climate
model projection.}
}
\value{
A list of length 2. First element is the name of the ensemble that was processed. Second element is
a list containing the coordinate .csv, the time series data .csv, and the time data .csv respectively.
}
\description{
List all files for a single ensemble member
}
\examples{
ensemble <- "/Users/brookeanderson/Downloads/sample/cmip5/rcp85/bcc1/r1i1p1"
buildStructureEnsembles(ensemble)
}

