% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{buildStructureExperiments}
\alias{buildStructureExperiments}
\title{Generate list of file structure for an experiment}
\usage{
buildStructureExperiments(model, experiment, all, dataPath, coordinateFilenames,
  tasFilenames, timeFilenames)
}
\arguments{
\item{model}{Character string of climate model name}

\item{experiment}{Character string of the experiment of interest.
Possible variables are "historical" or "rcp85".}

\item{all}{Character vectors with the relative pathnames of all
climate projection files of interest from the directory
specified in dataPath}

\item{dataPath}{Character string of the file path to the directory
containing the climate projections. Must include the final `/`.}
}
\value{
A list that is the length of the number of ensembles. Each element is a return value
of the buildStructureEnsembles function.
}
\description{
Generate list of file structure for an experiment
}
\examples{
model <- "bcc1"
experiment <- "rcp85"
all <- c("rcp85/bcc1/r1i1p1/latitude_longitude_NorthAmerica_12mo.csv",
         "rcp85/bcc1/r1i1p1/tas_NorthAmerica_12mo.csv",
         "rcp85/bcc1/r1i1p1/time_NorthAmerica_12mo.csv")
dataPath <- "~/Downloads/sample/cmip5/"
buildStructureExperiments(model, experiments, all, dataPath)

}

