% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{buildStructureExperiments}
\alias{buildStructureExperiments}
\title{Generate list of file structure for an experiment}
\usage{
buildStructureExperiments(model, experiment, dataFolder, coordinateFilenames,
  tasFilenames, timeFilenames)
}
\arguments{
\item{model}{Character string of climate model name}

\item{experiment}{Character string of the experiment of interest.
Possible variables are "historical" or "rcp85".}

\item{dataPath}{Character string of the file path to the directory
containing the climate projections. Must include the final `/`.}
}
\value{
A list that is the length of the number of ensembles. Each element is a return value
of the buildStructureEnsembles function.
}
\description{
Generate list of file structure for an experiment
}
\examples{
model <- "bcc1"
experiment <- "rcp85"

dataFolder <- system.file("cmip5", package = "futureheatwaves")
dataFolder <- paste0(dataFolder, "/")

coordinateFilenames <- "latitude_longitude_NorthAmerica_12mo.csv"
tasFilenames <- "tas_NorthAmerica_12mo.csv"
timeFilenames <- "time_NorthAmerica_12mo.csv"

buildStructureExperiments(model = model,
                          experiment = experiment,
                          dataFolder = dataFolder,
                          coordinateFilenames = coordinateFilenames,
                          tasFilenames = tasFilenames,
                          timeFilenames = timeFilenames)

}

