% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{buildStructureModels}
\alias{buildStructureModels}
\title{Generate list of file structure from model, experiment, and path}
\usage{
buildStructureModels(model, experiments, all, dataPath)
}
\arguments{
\item{model}{Character string of climate model name}

\item{experiments}{Character vector of the experiment(s) of interest.
Possible variables are "historical", "rcp85", or both.}

\item{all}{Character vectors with the relative pathnames of all
climate projection files of interest from the directory
specified in dataPath}

\item{dataPath}{Character string of the file path to the directory
containing the climate projections. Must include the final `\`.}
}
\description{
Generate list of file structure from model, experiment, and path
}
\examples{
model <- "bcc1"
experiments <- c("rcp85")
all <- c("rcp85/bcc1/r1i1p1/latitude_longitude_NorthAmerica_12mo.csv",
         "rcp85/bcc1/r1i1p1/tas_NorthAmerica_12mo.csv",
         "rcp85/bcc1/r1i1p1/time_NorthAmerica_12mo.csv")
dataPath <- "~/Downloads/sample/cmip5/"
buildStructureModels(model, experiments, all, dataPath)
}

