% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_params}
\alias{check_params}
\title{Error checking for parameters of gen_hw_set}
\usage{
check_params(out, dataFolder, citycsv, coordinateFilenames, tasFilenames,
  timeFilenames, RorCPP, IDheatwavesReplacement, dataBoundaries,
  referenceBoundaries)
}
\arguments{
\item{out}{Character string with pathway to directory to which
heatwave files will be written.}

\item{dataFolder}{Character string with pathway to directory that
contains climate projections. Must include the final backslash.}

\item{citycsv}{Character string giving the filepath to a .csv
file with latitude and longitude values for each city.}

\item{coordinateFilenames}{Character string with name of the files containing the latitude and longitude coordinates
corresponding to the columns of the time series data.}

\item{tasFilenames}{Character sting with name of files containing the time series data.}

\item{timeFilenames}{Character string with name of the files containing the date information corresponding,
to the rows of the time series data.}

\item{RorCPP}{0 /1 flag that indicates whether to use R (1) or
CPP (0) functions}

\item{IDheatwavesReplacement}{Either FALSE, to use the default
heatwave definition, or a user-specified custom function to
use to identify heatwaves.}

\item{dataBoundaries}{Custom time boundaries for extracting data
from the ensembles. Format: c(historical low bound, historical
high bound, reference low bound, reference high bound).
Restrictions: Bounds cannot span multiple experiments}

\item{referenceBoundaries}{Reference boundaries.}
}
\value{
Stops and returns an error if any parameters are incorrect.
}
\description{
Error checking for parameters of gen_hw_set
}
\note{
Does not check if the data is organized in the proper structure or
   if any data exists within the directory at all.

Does not check if the three ensemble .csv data files exist, only if they have
   the .csv extension.
}
\examples{
out <- "~/Downloads/sample/results"
dataFolder <- "~/Downloads/sample/cmip5/"
citycsv <- "inst/cities.csv"
referenceBoundaries <- FALSE
check_params(out, dataFolder, citycsv,
   referenceBoundaries = referenceBoundaries)
}

