% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_params}
\alias{check_params}
\title{Check for input parameter errors}
\usage{
check_params(out, dataFolder, dataDirectories, citycsv, coordinateFilenames,
  tasFilenames, timeFilenames, IDheatwavesFunction, thresholdBoundaries,
  projectionBoundaries, referenceBoundaries)
}
\arguments{
\item{out}{Character string with pathway to directory to which
heatwave files will be written. This should be a pathname to a directory
on the user's local computer. If the directory already exists, it will
be overwritten by this function, so the user should either specify a
pathname for a directory that does not yet exist.}

\item{dataFolder}{Character string with pathway to directory that
contains climate projection data.}

\item{citycsv}{Character string giving the filepath to a
comma-separated values (.csv) file with unique identifiers, latitudes, and
longitudes for each community for which the user wants to generate
heatwave projections.}

\item{coordinateFilenames}{Character string with filename of the file
containing the latitude and longitude coordinates in the bottom
directory of each ensemble member subdirectory. (See the package
vignette for an example of the required structure for this file.)}

\item{tasFilenames}{Character string with filename of the file
containing the climate projection data in the bottom
directory of each ensemble member subdirectory. (See the package
vignette for an example of the required structure for this file.)}

\item{timeFilenames}{Character string with filename of the file
containing the date dataframe in the bottom
directory of each ensemble member subdirectory. (See the package
vignette for an example of the required structure for this file.)}

\item{IDheatwavesFunction}{A character string with the name of the R function
to use to identify heatwaves. This function may be a user-specified custom
function to, but it must be loaded into the current R session. The
function name must be put in quotation marks.}

\item{thresholdBoundaries}{A numeric vector with the custom time boundaries
to be used to determine the threshold temperatures for the heatwave
definition. The required format for this vector is that the first element
is the lower year bound and the second element is the upper year bound,
with the restrictions that bounds must be between the years 1980 and 2099
and cannot span 2005 (i.e., both must either be before 2005 or after 2005).}

\item{projectionBoundaries}{A numeric vector with the custom time boundaries
for which the user wants to create heatwave projections. The required
format for this vector is that the first element
is the lower year bound and the second element is the upper year bound,
with the restrictions that bounds must be between the years 1980 and 2099
and cannot span 2005 (i.e., both must either be before 2005 or after 2005).}

\item{referenceBoundaries}{A numeric vector with the custom time boundaries
to use in calculating relative characteristics for heatwaves (i.e., to use
when exploring the role of adaptation in projections). The required format
for this vector is that the first element is the lower year bound and the
second element is the upper year bound, with the restrictions that bounds
must be between the years 1980 and 2099 and cannot span 2005 (i.e., both
must either be before 2005 or after 2005).}
}
\value{
Stops and returns an error if any parameters are incorrect.
}
\description{
This function goes through all parameter inputs for the main
functions, \code{\link{gen_hw_set}}, and makes sure all parameter
entries are in the appropriate format for following functions.
If any parameters are in an incorrect format, the function stops
and returns an error describing the problem.
}
\note{
This function does not check if the data is organized in the proper
   structure or if any data exists within the directory at all, so a
   call to \code{\link{gen_hw_set}} could still pass through this check and
   make it further through the function code with those mistakes.

 Does not check if the three ensemble final .csv data files exist,
   only if they have the .csv extension if they do exist. (Reminder:
   the final subdirectory should have the following three csv files:
   1. A file with the climate model projections, with grid points by
   column and times by row; 2. A file with the longitude and latitude
   of each grid point in the projection file and; 3. A file with the
   date of each of the rows in the projection file.)
}

