% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{checkCustomBounds}
\alias{checkCustomBounds}
\title{Error checking for user-specified custom boundaries}
\usage{
checkCustomBounds(boundList, length = 4)
}
\arguments{
\item{boundList:}{A set of boundaries in the format c(historical period low bound, historical period high bound,
rcp period low bound, rcp period high bound)}

\item{Length:}{The required length of the boundaries list. Defaulted to 4.}
}
\description{
Error checking for user-specified custom boundaries
}

