% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{checkCustomBounds}
\alias{checkCustomBounds}
\title{Check year boundaries for errors}
\usage{
checkCustomBounds(boundList)
}
\arguments{
\item{boundList}{Either FALSE, if user wishes to use the default
boundaries, or a set of boundary years in the format
c(historical period earliest year, historical period latest year,
future projection period earliest year, future projection period
latest year).}
}
\description{
This function inputs the boundary lists specified in
\code{\link{gen_hw_set}}, \code{dataBoundaries} and
\code{referenceBoundaries}, and checks them for errors in structure of the
input or in the years selected.
}

