% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{closest_point}
\alias{closest_point}
\title{Find closest grid point to a city location}
\usage{
closest_point(city, latlong)
}
\arguments{
\item{city}{A numeric vector containing the city's ID, latitude, and
longitude, in that order.}

\item{latlong}{A dataframe of latitude and longitude pairings for each
climate grid cell location, with latitudes in the first column and
longitudes in the second column.}
}
\value{
An index corresponding to the coordinates in the dataframe
   that is closest to the city coordinates.
}
\description{
Identify the closest grid points for a climate model to a
city based on the city's latitude and longitude. Use the
Pythagorean theorem.
}

