% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{closest_point}
\alias{closest_point}
\title{Find closest grid point to a city location}
\usage{
closest_point(city, latlong)
}
\arguments{
\item{city}{A numeric vector containing the city's ID, latitude, and
longitude, in that order.}

\item{latlong}{A dataframe of latitude and longitude of each
climate grid cell, with latitudes in the first column and
longitudes in the second column. The \code{\link{readLatLong}} function
re-orders the columns of longitude and latitude to comply with this
format if necessary.}
}
\value{
An index corresponding to the column in the climate projections
   dataframe for that climate model that is closest to the city's
   coordinates.
}
\description{
This function identifies the closest grid points in a climate model to a
city based on the city's latitude and longitude using the
Euclidean distance.
}

