% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHeatwaves.R
\name{consolidate}
\alias{consolidate}
\title{Combine all identified heatwave dataframes together into one.}
\usage{
consolidate(hwDataframeList)
}
\arguments{
\item{hwDataframeList}{A list object where each element is the dataframe
of identified and characterized heatwaves, created by
the closure created by \code{\link{createCityProcessor}},
for a single community .}
}
\value{
A combined dataframe version of the list that was passed as an
   argument.
}
\description{
Combine all identified heatwave dataframes together into one.
}

