% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHeatwaves.R
\name{consolidate}
\alias{consolidate}
\title{Combine all identified heatwave dataframes together into one.}
\usage{
consolidate(hwDataframeList)
}
\arguments{
\item{hwDataframeList}{A list of identified heatwave dataframes created by
the closure of createCityProcessor}
}
\value{
A combined dataframe of identified heatwave dataframes from the
   dataframe list that was passed as an argument.
}
\description{
Combine all identified heatwave dataframes together into one.
}

