% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHeatwaves.R
\name{consolidate}
\alias{consolidate}
\title{Consolidate heatwave dataframes}
\usage{
consolidate(hwDataframeList)
}
\arguments{
\item{hwDataframeList}{A list object where each element is the dataframe
of heatwaves, created by the closure created by
\code{\link{createCityProcessor}}, for a single city.}
}
\value{
A combined dataframe version of the list object that was passed as an
   argument.
}
\description{
This function combines all identified city-specific heatwave dataframes
together into a single dataframe. This function is used to create a single
dataframe with all heatwaves from all study cities for an ensemble
member.
}

