% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHeatwaves.R
\name{consolidate}
\alias{consolidate}
\title{Consolidate heatwave dataframes}
\usage{
consolidate(hwDataframeList)
}
\arguments{
\item{hwDataframeList}{A list object where each element is the dataframe
of identified and characterized heatwaves, created by
the closure created by \code{\link{createCityProcessor}},
for a single community .}
}
\value{
A combined dataframe version of the list that was passed as an
   argument.
}
\description{
This function combines all identified heatwave dataframes together into
a single dataframe. This function is used to create a single dataframe
with all heatwaves from all specified communities for each ensemble
member.
}

