% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHeatwaves.R
\name{createCityProcessor}
\alias{createCityProcessor}
\title{Create closure to identify and aggregate heatwaves}
\usage{
createCityProcessor(global)
}
\arguments{
\item{global}{An object name for the \code{global} object. This
object is created by \code{\link{gen_hw_set}} and includes user
specifications for the path to the output directory, the path to the
input climate projections, the dataframe with community locations, and the
format for the filenames of the coordinate, tas, and time files in the
ensemble-specific subdirectories of the projection directory input by
the user.}
}
\value{
This function creates a closure that takes inputs of \code{threshold},
   \code{ensembleSeries}, and \code{custom} and returns a closure that will
   find and characterize all heatwaves in all communities for a given
   ensemble. See the help file for \code{\link{formHwFrame}} for more
   information on the format of the dataframe created by this closure.
}
\description{
This function creates a closure that returns a
   dataframe with the identified heatwaves and heatwave characteristics for
   a given city for the specified projection period, as generated by the
   \code{\link{createHwDataframe}} function.
}
\note{
The closure encapsulates an incrementer varaible and advances it
   with every call. This variable is used to index into the \code{cities}
   vector from the \code{global} object passed into this function.
}

