% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHeatwaves.R
\name{createCityProcessor}
\alias{createCityProcessor}
\title{Identify and aggregate heatwaves for a particular city}
\usage{
createCityProcessor(global, custom)
}
\arguments{
\item{global}{List of global parameters generated by listGlobal.}

\item{custom}{List of custom parameters generated by listCustom.}
}
\value{
A closure that will find all heatwaves for a given ensemble. Each passed threshold corresponds to a particular
city. The closure advances an incrementer every time it is called to get this effect.
Argument 1: A threshold temperature
Argument 2: The ensemble from which you wish to find heatwaves.
}
\description{
Identify and aggregate heatwaves for a particular city
}
\note{
The closure encapsulates an incrementer varaible and advances it with every call. This variable is used to
index into the cities vector from the "global" data list.
}

