% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHeatwaves.R
\name{createCityProcessor}
\alias{createCityProcessor}
\title{Identify and aggregate heatwaves for a particular city}
\usage{
createCityProcessor(global, custom)
}
\arguments{
\item{global}{List of global parameters generated by listGlobal.}

\item{custom}{List of custom parameters generated by listCustom.}
}
\value{
List with all heatwaves for the cities of interest
   for a climate projection. (is this right?)
}
\description{
Identify and aggregate heatwaves for a particular city
}

