% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{createEnsembleWriter}
\alias{createEnsembleWriter}
\title{Ensemble writer factory function}
\usage{
createEnsembleWriter(modelName, global, custom)
}
\arguments{
\item{modelName}{Name of the model}

\item{global}{The global data list}

\item{custom}{The custom parameter list}
}
\value{
A closure that writes heatwave dataframes
Argument 1: A combined heatwave dataframe that contains all heatwave
information for the ensemble being processed.

#' @importFrom dplyr %>%
}
\description{
Produces a function that writes a single heatwave list to a .csv
}
\details{
Closes over an incrementer variable for ensembles that advances forward
each time the produced function is called.
Output the heatwave data of each model. Produces a closure that must be
applied to a list of the heatwave dataframes corresponding to the
ensembles of the model
}

