% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{createEnsembleWriter}
\alias{createEnsembleWriter}
\title{Ensemble writer factory function}
\usage{
createEnsembleWriter(modelName, global, custom)
}
\arguments{
\item{modelName}{Character string of climate model name (e.g., "bcc1"). This
name is typically generated for use in this function from the subdirectory
names for the climate model within the directory of projection data
specified by the user in \code{\link{gen_hw_set}}.}

\item{global}{An object name for the 'global' variable list. This
parameter is passed through from \code{\link{gen_hw_set}} and includes user
specifications for the path to the output directory, the path to the
input climate projections, the dataframe with community locations, the
format for the filenames of the coordinate, tas, and time files in the
ensemble-specific subdirectories of the projection directory input by
the user, and whether the user prefers to run the function using the R
or C++ version of the function to identify heatwaves.}

\item{custom}{An object name for the 'custom` data list. This parameter is
passed through from \code{\link{gen_hw_set}} and includes user specifications
for, if specified as other than the default, an alternative function to
use to identify heatwaves, alternative upper and lower year boundaries
for the projection period of the heatwave datasets being generated,
alternative upper and lower year boundaries for the reference period
to be used when measuring all heatwave characteristics related to
relative, rather than absolute, temperature, and the percentile
threshold to use when defining heatwaves.}
}
\value{
A closure that writes heatwave dataframes
Argument 1: A combined heatwave dataframe that contains all heatwave
information for the ensemble being processed.
}
\description{
Produces a function that writes a single heatwave list to a .csv
}
\details{
Closes over an incrementer variable for ensembles that advances forward
each time the produced function is called.
Output the heatwave data of each model. Produces a closure that must be
applied to a list of the heatwave dataframes corresponding to the
ensembles of the model
}

