% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHeatwaves.R
\name{createHwDataframe}
\alias{createHwDataframe}
\title{Characterize heatwaves}
\usage{
createHwDataframe(city, threshold, heatwaves, ensembleSeries, i, global, custom)
}
\arguments{
\item{threshold}{Numeric string with threshold temperature used in
the heatwave definition, in the same units as the temperature time
series in \code{datafr}.}

\item{heatwaves}{data.frame(dates, thresholds)}

\item{ensembleSeries}{A list object giving the projection times series as
well as a variety of other information for a single ensemble member.
This is the output of \code{\link{processEnsemble}}.}

\item{i}{An index specifying with community of the specified communities
is being processed. This corresponds to the order that the communities
are given in the \code{citycsv} file specified in
\code{\link{gen_hw_set}}.}

\item{global}{An object name for the 'global' variable list. This
parameter is passed through from \code{\link{gen_hw_set}} and includes user
specifications for the path to the output directory, the path to the
input climate projections, the dataframe with community locations, and the
format for the filenames of the coordinate, tas, and time files in the
ensemble-specific subdirectories of the projection directory input by
the user.}

\item{custom}{An object name for the 'custom` data list. This parameter is
passed through from \code{\link{gen_hw_set}} and includes user specifications
for, if specified as other than the default, the name of the R function to
use to identify heatwaves, alternative upper and lower year boundaries
for the data used to determine threshold temperatures for the heatwave
definition, alternative upper and lower year boundaries
for the projection period of the heatwave datasets being generated,
alternative upper and lower year boundaries for the reference period
to be used when measuring all heatwave characteristics related to
relative, rather than absolute, temperature, and the percentile
threshold to use when defining heatwaves.}
}
\value{
Result is a dataframe where each row represents a heatwave.
}
\description{
Characterize heatwaves
}

