% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHeatwaveDataframe.R
\name{createHwDataframe}
\alias{createHwDataframe}
\title{Aggregate by heatwave}
\usage{
createHwDataframe(city = stop("Unspecified city"),
  threshold = stop("Unspecified threshold"),
  heatwaves = stop("'heatwaves' unspecified"), percentile = 0.98, ensemble,
  i, global, custom)
}
\arguments{
\item{city}{Index of cities global}

\item{threshold}{Threshold temperature for this city}

\item{heatwaves}{data.frame(dates, thresholds)}

\item{percentile}{Percentile for threshold)}
}
\value{
Result is a dataframe where each row represents a heatwave.
}
\description{
Aggregate by heatwave
}

