% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHeatwaves.R
\name{formHwFrame}
\alias{formHwFrame}
\title{Create heatwave dataframe for the given ensemble}
\usage{
formHwFrame(ensemble, thresholds, global, custom)
}
\arguments{
\item{ensemble}{Character string that gives the absolute file path
for the directory with a particular ensemble member of a climate
model projection.}

\item{thresholds}{[What is this?]}

\item{global}{List of global parameters generated by listGlobal.}

\item{custom}{List of custom parameters generated by listCustom.}
}
\value{
The combined dataframe of identified heatwaves for all cities of the given ensemble.
}
\description{
Create heatwave dataframe for the given ensemble
}

