% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futureheatwaves.R
\docType{package}
\name{futureheatwaves}
\alias{futureheatwaves}
\alias{futureheatwaves-package}
\title{Find, Characterize, and Explore Heat Waves in Climate Projections}
\description{
\code{futureheatwaves} takes a directory of climate projection files and,
for each, identifies and characterizes all heat waves. The definition used to
identify heat waves can be customized. Characterizations include several
metrics of heat wave length, intensity, and timing in the year. Heat waves
can be explored by applying custom functions across all generated heat wave
files. This work was supported in part by grants from the National Institute
of Environmental Health Sciences (R00ES022631), the National Science
Foundation (1331399), and the Colorado State University Vice President for
Research.
}

