% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gen_hw_set}
\alias{gen_hw_set}
\title{Create and write heatwave projections}
\usage{
gen_hw_set(out, dataFolder, dataDirectories = list(historical = c(1980, 2004),
  rcp85 = c(2006, 2099)), citycsv, coordinateFilenames, tasFilenames,
  timeFilenames, IDheatwavesFunction = "IDHeatwavesCPPwrapper",
  thresholdBoundaries = c(1981, 2004), projectionBoundaries = c(2061, 2080),
  referenceBoundaries = c(2061, 2080), models_to_run = "all",
  probThreshold = 0.98, printWarning = TRUE,
  threshold_ensemble = "r1i1p1", lat_lon_colnames = c("lat", "lon"))
}
\arguments{
\item{out}{Character string with pathway to directory to which
heatwave files will be written. This should be a pathname to a directory
on the user's local computer. If the directory already exists, it will
be overwritten by this function, so the user should either specify a
pathname for a directory that does not yet exist.}

\item{dataFolder}{Character string with pathway to directory that
contains climate projection data.}

\item{dataDirectories}{A list object, with two elements for each of the
two subdirectories included in the main directory. Typically, these will
be a historical directory and a climate projection directory. Each element
of the list should be named with the names of the subdirectories and
should provide a numeric vector with the starting year and ending year of
the data within each of the two subdirectories (e.g.,
\code{list("historical" = c(1980, 2004), "rcp85" = c(2006, 2099))})}

\item{citycsv}{Character string giving the filepath to a
comma-separated values (.csv) file with unique identifiers, latitudes, and
longitudes for each community for which the user wants to generate
heatwave projections.}

\item{coordinateFilenames}{Character string with filename of the file
containing the latitude and longitude coordinates in the bottom
directory of each ensemble member subdirectory. (See the package
vignette for an example of the required structure for this file.)}

\item{tasFilenames}{Character string with filename of the file
containing the climate projection data in the bottom
directory of each ensemble member subdirectory. (See the package
vignette for an example of the required structure for this file.)}

\item{timeFilenames}{Character string with filename of the file
containing the date dataframe in the bottom
directory of each ensemble member subdirectory. (See the package
vignette for an example of the required structure for this file.)}

\item{IDheatwavesFunction}{A character string with the name of the R function
to use to identify heatwaves. This function may be a user-specified custom
function to, but it must be loaded into the current R session. The
function name must be put in quotation marks.}

\item{thresholdBoundaries}{A numeric vector with the custom time boundaries
to be used to determine the threshold temperatures for the heatwave
definition. The required format for this vector is that the first element
is the lower year bound and the second element is the upper year bound,
with the restrictions that bounds must be between the years 1980 and 2099
and cannot span 2005 (i.e., both must either be before 2005 or after 2005).}

\item{projectionBoundaries}{A numeric vector with the custom time boundaries
for which the user wants to create heatwave projections. The required
format for this vector is that the first element
is the lower year bound and the second element is the upper year bound,
with the restrictions that bounds must be between the years 1980 and 2099
and cannot span 2005 (i.e., both must either be before 2005 or after 2005).}

\item{referenceBoundaries}{A numeric vector with the custom time boundaries
to use in calculating relative characteristics for heatwaves (i.e., to use
when exploring the role of adaptation in projections). The required format
for this vector is that the first element is the lower year bound and the
second element is the upper year bound, with the restrictions that bounds
must be between the years 1980 and 2099 and cannot span 2005 (i.e., both
must either be before 2005 or after 2005).}

\item{models_to_run}{A character vector with either "all" (the default),
in which case the function runs through all models in the specified data
directory, or the names of the models to run, using the names of each
model's subdirectory within the data directory (e.g.,
\code{c("bcc1", "ccsm")})}

\item{probThreshold}{Numerical value between 0 and 1 specifying the threshold
of temperature to use when defining heatwaves. The default value is 0.98
(i.e., a heatwave is two or more days above the community's 98th
percentile temperature).}

\item{printWarning}{TRUE / FALSE, specifies whether to print out a warning
informing the user that the function will write out results to the local
directory specified by the user with \code{out}. This warning prints out
by default; the user must opt-out of this warning by specifying FALSE
for this argument.}

\item{threshold_ensemble}{A character vector giving the name of the ensemble
member that should be used when determining the threshold temperature for
the heatwave definition for each community for each climate model (e.g.,
\code{"r1i1p1"})}

\item{lat_lon_colnames}{A character vector of length two with the column names
in the \code{cities} dataframe for latitude (first vector element) and
longitude (second vector element)}
}
\value{
This function returns a dataframe listing the name of each climate
   model used, as well as the number of historical and future projection
   ensembles for each model. This output can be used as a check
   that the function processed through the directory of input files
   specified using the \code{dataFolder} argument.

This function also creates, and writes to the user's computer, files with
   the heatwaves and their characteristics for the specified climate
   projections and dates.
}
\description{
This function creates heatwave projection datasets for all models and
ensemble members for a user-specified set of communities.The resulting
heatwave projections are written out to a specified directory on the
user's local computer.
}

