% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gen_hw_set}
\alias{gen_hw_set}
\title{Create and write heatwave projections}
\usage{
gen_hw_set(out, dataFolder, citycsv, coordinateFilenames, tasFilenames,
  timeFilenames, RorCPP = 1, IDheatwavesReplacement = FALSE,
  dataBoundaries = FALSE, referenceBoundaries = FALSE,
  probThreshold = 0.98, printWarning = TRUE)
}
\arguments{
\item{out}{Character string with pathway to directory to which
heatwave files will be written.}

\item{dataFolder}{Character string with pathway to directory that
contains climate projection data.}

\item{citycsv}{Character string giving the filepath to a .csv
file with latitude and longitude values for each city.}

\item{coordinateFilenames}{Character string with name of the files
containing the latitude and longitude coordinates
corresponding to the columns of the time series data.}

\item{tasFilenames}{Character sting with name of files containing the time
series data.}

\item{timeFilenames}{Character string with name of the files containing the
date information corresponding to the rows of the time series data.}

\item{RorCPP}{0 /1 flag that indicates whether to use R (0) or
C++ (1) function to identify heatwaves in the projections}

\item{IDheatwavesReplacement}{Either FALSE, to use the default
heatwave definition, or a user-specified custom function to
use to identify heatwaves.}

\item{dataBoundaries}{Custom time boundaries for extracting data
from the ensembles. Format: c(historical low bound, historical
high bound, reference low bound, reference high bound).
Restrictions: Bounds cannot span multiple experiments}

\item{referenceBoundaries}{Custom time boundaries for ... .
The first two values give the year range of the dataset to use to
determine the threshold temperatures of the heatwave definition.
Format: c(historical low bound, historical
high bound, reference low bound, reference high bound).
Restrictions: Bounds cannot span multiple experiments}

\item{probThreshold}{Numerical value between 0 and 1 specifying the threshold
of temperature to use when defining heatwaves. The default value is 0.98
(i.e., a heatwave is two or more days above the community's 98th
percentile temperature).}

\item{printWarning}{TRUE / FALSE specifying whether to print out the progress
of the function as it runs. Default is TRUE.}
}
\value{
This function returns a dataframe listing the name of each climate
   model included in the directory of projection files inputted to the
   function as well as the number of historical and future projection
   ensembles included for each model. This output can be used as a check
   that the function processed through the directory of input files
   specified using the \code{dataFolder} argument.
}
\description{
This function creates heatwave projection datasets for all models and
   ensemble members in a directory of climate projections for a specified
   set of communities.The resulting heatwave projections are written out
   to a specified directory on the user's local computer.
}

