% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{getBounds}
\alias{getBounds}
\title{end' index of the historical experiment has to be matched from back to front.}
\usage{
getBounds(times, experiment, custom)
}
\description{
end' index of the historical experiment has to be matched from back to front.
}

