% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{getBounds}
\alias{getBounds}
\title{Acquire boundaries of time series data}
\usage{
getBounds(times, experiment, custom)
}
\arguments{
\item{times}{A dataframe containing the time data for one of the
climate model ensemble's projection. Read in by \code{processEnsemble}.}

\item{experiment}{Character string of the experiment of
interest. Possible variables are "historical" or "rcp85".}

\item{custom}{An object name for the 'custom` data list. This parameter is
passed through from \code{gen_hw_set} and includes user specifications
for, if specified as other than the default, an alternative function to
use to identify heatwaves, alternative upper and lower year boundaries
for the projection period of the heatwave datasets being generated,
alternative upper and lower year boundaries for the reference period
to be used when measuring all heatwave characteristics related to
relative, rather than absolute, temperature, and the percentile
threshold to use when defining heatwaves.}
}
\value{
A numeric vector containing an upper bound of the experiment
   time period as an index; the lower bound of the experiment time
   period as an index; and the number of days spanned.
   Format: c(start, end, size)
}
\description{
Acquire boundaries of time series data
}
\note{
This function is customizable.
end' index of the historical experiment has to be matched from back to front.
}

