% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{getBounds}
\alias{getBounds}
\title{Acquire boundaries of time series data}
\usage{
getBounds(times, custom, type)
}
\arguments{
\item{times}{A dataframe containing the time data for one of the
climate model ensemble's projection. Each row in this dataframe
corresponds to the row with the same row number in the climate
projection dataframe. Read in by
\code{\link{processEnsemble}}.}

\item{custom}{An object name for the 'custom` data list. This parameter is
passed through from \code{\link{gen_hw_set}} and includes user specifications
for, if specified as other than the default, an alternative function to
use to identify heatwaves, alternative upper and lower year boundaries
for the projection period of the heatwave datasets being generated,
alternative upper and lower year boundaries for the reference period
to be used when measuring all heatwave characteristics related to
relative, rather than absolute, temperature, and the percentile
threshold to use when defining heatwaves.}
}
\value{
A numeric vector containing the number of days spanned by the
   time period, the lower bound of the experiment time
   period as an index specifying the relevant row of the climate projection
   data; and an upper bound of the experiment time period as an index
   specifying the relevant row of the climate projection data.
}
\description{
Acquire boundaries of time series data
}
\note{
This function is customizable. If data boundaries are not set
   using \code{dataBoundaries} in the \code{\link{gen_hw_set}} function call,
   the default is to pull Jan. 1, 1981, to Dec. 31, 2014 for a
   historical experiment and Jan. 1, 2061 to Dec. 31, 2080 for
   a future projection experiment.
}

