% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleApplyFunctions.R
\name{heatwave_days}
\alias{heatwave_days}
\title{Calculate total heat wave days}
\usage{
heatwave_days(hw_datafr)
}
\arguments{
\item{hw_datafr}{A dataframe of heat waves and their characteristics, as
created by \code{\link{gen_hw_set}}}
}
\value{
A numeric value with the number of heat wave days in a dataframe
   of heat waves, as generated by \code{\link{gen_hw_set}}.
}
\description{
This function takes a dataframe of heat waves, as created by
\code{\link{gen_hw_set}}, and calculates the total number of heat wave days in
the dataframe (sum of the number of days in each heat wave for all heat waves
in the dataset).
}
\note{
This function is an example of a function that can be created and
   used to explore heat waves using the \code{\link{apply_all_models}}
   function.
}
\examples{
data(hw_datafr)
heatwave_days(hw_datafr)

}

