% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDHeatwaves.R
\name{IDheatwaves}
\alias{IDheatwaves}
\title{Identify all heatwaves within a given city's time series}
\usage{
IDheatwaves(city, threshold, days = 2, datafr, global, custom)
}
\arguments{
\item{city}{A numeric vector containing the city's ID, latitude, and
longitude, in that order.}

\item{threshold}{Numeric string with threshold percentile used in
heatwave definition.}

\item{days}{Numeric string specifying minimum number of days required
in heatwave definition.}

\item{datafr}{A dataframe with daily temperature projections in the
the community being processed. This dataframe must have two columns:
(1) the first column must have the date of each observation, with
class "Date" and; (2) the second column must have temperatures, in
Fahrenheit. In the normal running of this package, this dataframe
will be generate by the closure created by
\code{\link{createCityProcessor}}.}

\item{global}{An object name for the 'global' variable list. This
parameter is passed through from \code{\link{gen_hw_set}} and includes user
specifications for the path to the output directory, the path to the
input climate projections, the dataframe with community locations, the
format for the filenames of the coordinate, tas, and time files in the
ensemble-specific subdirectories of the projection directory input by
the user, and whether the user prefers to run the function using the R
or C++ version of the function to identify heatwaves.}

\item{custom}{An object name for the 'custom` data list. This parameter is
passed through from \code{\link{gen_hw_set}} and includes user specifications
for, if specified as other than the default, an alternative function to
use to identify heatwaves, alternative upper and lower year boundaries
for the projection period of the heatwave datasets being generated,
alternative upper and lower year boundaries for the reference period
to be used when measuring all heatwave characteristics related to
relative, rather than absolute, temperature, and the percentile
threshold to use when defining heatwaves.}
}
\value{
Returns the dataframe entered as \code{datafr}, but with new
   columns providing heatwave identifiers. The returned dataframe will
   have new columns for whether a day was part of a heatwave (\code{hw},
   0 / 1), if it was part of a heatwave, the number of the heatwave
   (\code{hw.number}), and whether the day was the first day in a heatwave
   (\code{first.hw.day}, 0 /1).
}
\description{
Identify all heatwaves within a given city's time series
}

