% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDHeatwavesForCountry.R
\name{IDheatwaves}
\alias{IDheatwaves}
\title{Identify all heatwaves within a given city's time series}
\usage{
IDheatwaves(city, threshold, days = 2, datafr, global, custom)
}
\arguments{
\item{city}{Character string with city ID.}

\item{threshold}{Numeric string with threshold percentile used in
heatwave definition.}

\item{days}{Numeric string specifying minimum number of days required
in heatwave definition.}

\item{datafr}{Dataframe with daily temperatures in city of interest.}

\item{global}{[Global holder-- what's that?]}

\item{custom}{[Custom data holder-- what's that?]}
}
\value{
Returns the dataframe entered as \code{datafr}, but with new
   columns providing heatwave idendifiers.
}
\description{
Identify all heatwaves within a given city's time series
}

