% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDHeatwaves.R
\name{IDHeatwavesAlternative}
\alias{IDHeatwavesAlternative}
\title{Identify heatwaves with alternative definition}
\usage{
IDHeatwavesAlternative(city = stop("Error: unspecified city"),
  threshold = stop("Error: unspecified threshold"), days = 1,
  datafr = stop("Error: 'datafr' unspecified"))
}
\arguments{
\item{city}{...}

\item{threshold}{Numeric string with threshold percentile used in
heatwave definition.}

\item{days}{Numeric string specifying minimum number of days required
in heatwave definition.}

\item{datafr}{Dataframe with a time series of the daily temperatures in
the community being analyzed.}
}
\value{
Returns the dataframe entered as \code{datafr}, but with new
   columns providing heatwave identifiers.
}
\description{
Identify heatwaves with alternative definition
}

