% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDHeatwaves.R
\name{IDHeatwavesAlternative}
\alias{IDHeatwavesAlternative}
\title{Identify heatwaves with alternative definition}
\usage{
IDHeatwavesAlternative(threshold = stop("Error: unspecified threshold"),
  days = 5, datafr = stop("Error: 'datafr' unspecified"))
}
\arguments{
\item{threshold}{Numeric string with threshold percentile used in
heatwave definition.}

\item{days}{Numeric string specifying minimum number of days required
in heatwave definition.}

\item{datafr}{A dataframe with daily temperature projections in the
the community being processed. This dataframe must have two columns:
(1) the first column must have the date of each observation, with
class "Date" and; (2) the second column must have temperatures, in
Fahrenheit. In the normal running of this package, this dataframe
will be generate by the closure created by
\code{\link{createCityProcessor}}.}
}
\value{
Returns the dataframe entered as \code{datafr}, but with new
   columns providing heatwave identifiers.
}
\description{
Identify heatwaves with alternative definition
}

