% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDHeatwavesForCountry.R
\name{IDHeatwavesAlternative}
\alias{IDHeatwavesAlternative}
\title{Function to identify heatwaves with alternative definition}
\usage{
IDHeatwavesAlternative(city = stop("Error: unspecified city"),
  threshold = stop("Error: unspecified threshold"), days = 1,
  datafr = stop("Error: 'datafr' unspecified"))
}
\arguments{
\item{city}{Character string with community identifies, as set in the
\code{citycsv} file specified by the user in \code{\link{gen_hw_set}}}

\item{threshold}{Numeric string with threshold percentile used in
heatwave definition.}

\item{days}{Numeric string specifying minimum number of days required
in heatwave definition.}

\item{datafr}{Dataframe with a time series of the daily temperatures in
the community being analyzed.}
}
\description{
Function to identify heatwaves with alternative definition
}

