% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDHeatwaves.R
\name{IDHeatwavesCPPwrapper}
\alias{IDHeatwavesCPPwrapper}
\title{Identify heatwaves in a time series}
\usage{
IDHeatwavesCPPwrapper(datafr, threshold)
}
\arguments{
\item{datafr}{A dataframe with daily temperature projections in the
the community being processed. This dataframe must have two columns:
(1) the first column must have the date of each observation, with
class "Date" and; (2) the second column must have temperatures
in degrees Fahrenheit. In the normal running of this package, this
dataframe will be generated by the closure created by
\code{\link{createCityProcessor}}.}

\item{threshold}{Numeric string with threshold temperature used in
the heatwave definition, in degrees Fahrenheit.}
}
\value{
Returns the dataframe entered as \code{datafr}, but with new
   columns providing heatwave identifiers. The returned dataframe will
   have new columns for whether a day was part of a heatwave (\code{hw},
   0 / 1), and, if it was part of a heatwave, the number of the heatwave
   (\code{hw.number}).
}
\description{
This function identifies heatwaves in a time series of temperature
data using a heatwave definition that a heatwave must be two or more
days with temperatures equal to or above some threshold temperature. This
function uses a compiled C++ function for part of the process, making it
faster than the R analogue, \code{\link{IDHeatwavesR}}, although the two
functions give identical results.
}
\note{
There are a few cases near the edges of data frames when this function
   would return that a day was not a heatwave when it was. First, if the
   first day of the dataset is a heatwave because preceeding days exceeded
   the threshold, but the second day in the dataframe is not above the
   threshold, this function would not capture that the first day was a
   heatwave. A similar caveat applies to the last day in the dataframe.
   In northern hemisphere studies, this should not be a concern, as it is
   unlikely that Jan. 1 or Dec. 31 would qualify as
   a heatwave. However, care should be taken
   when using this function either with Southern Hemisphere communities
   or when exploring exposures that, unlike heatwaves, may occur very
   early or late in the calendar year.
}

