% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDHeatwaves.R
\name{IDHeatwavesR}
\alias{IDHeatwavesR}
\title{Identify heatwaves in a time series}
\usage{
IDHeatwavesR(threshold, datafr)
}
\arguments{
\item{threshold}{Numeric string with threshold temperature used in
the heatwave definition, in the same units as the temperature time
series in \code{datafr}.}

\item{datafr}{A dataframe with daily temperature projections in the
the community being processed. This dataframe must have two columns:
(1) the first column must have the date of each observation, with
class "Date" and; (2) the second column must have temperatures
(in typical runs of this function, this will be in degrees
Fahrenheit). In the normal running of this package, this dataframe
will be generate by the closure created by
\code{\link{createCityProcessor}}.}
}
\value{
Returns the dataframe entered as \code{datafr}, but with new
   columns providing heatwave identifiers. The returned dataframe will
   have new columns for whether a day was part of a heatwave (\code{hw},
   0 / 1), if it was part of a heatwave, the number of the heatwave
   (\code{hw.number}), and whether the day was the first day in a heatwave
   (\code{first.hw.day}, 0 /1).
}
\description{
This function identifies heatwaves in a time series of temperature
data using a heatwave definition that a heatwave must be two or more
days with temperatures equal to or above some threshold temperature.
}
\details{
This function is the default function used to identify heatwaves in
\code{\link{gen_hw_set}}.
}
\note{
There are a few cases near the edges of data frames when this function
   would return that a day was not a heatwave when it was.First, if the first
   day of the dataset is a heatwave because preceeding days exceeded the
   threshold, but the second day in the dataframe is not above the threshold,
   this function would not capture that the first day was a heatwaves.
   Similar caveats apply to the last day in the dataframe. In northern
   hemisphere communities, this should not be a concern when studying
   heatwaves, as it is unlikely that Jan. 1 or Dec. 31 would qualify as
   a heatwave in this part of the world. However, care should be taken
   when using this function either with Southern Hemisphere communities
   or when exploring exposures that, unlike heatwaves, may occur very
   early or late in the calendar year.
}

