% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCities.R
\name{map_grid}
\alias{map_grid}
\title{Create a map of model grid}
\usage{
map_grid(plot_model, out)
}
\arguments{
\item{plot_model}{A character string with the name of the model to plot}

\item{out}{Character string with pathway to directory to which
heat wave files will be written. This should be a pathname to a directory
on the user's local computer. If the directory already exists, it will
be overwritten by this function, so the user should either specify a
pathname for a directory that does not yet exist.}
}
\description{
Create a map of model grid
}
\examples{
out <- system.file(package = "futureheatwaves")
map_grid(plot_model = "bcc1", out = out)

}

