% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processEnsemble}
\alias{processEnsemble}
\title{Extract projections from ensemble member}
\usage{
processEnsemble(ensemble, experiment, modelName, global, custom, reference)
}
\arguments{
\item{experiment}{Character string of the experiment of interest.
Possible variables are "historical" or "rcp85".}

\item{modelName}{Character string of climate model name (e.g., "bcc1"). This
name is typically generated for use in this function from the subdirectory
names for the climate model within the directory of projection data
specified by the user in \code{gen_hw_set}.}

\item{global}{An object name for the 'global' variable list. This
parameter is passed through from \code{gen_hw_set} and includes user
specifications for the path to the output directory, the path to the
input climate projections, the dataframe with community locations, the
format for the filenames of the coordinate, tas, and time files in the
ensemble-specific subdirectories of the projection directory input by
the user, and whether the user prefers to run the function using the R
or C++ version of the function to identify heatwaves.}

\item{custom}{An object name for the 'custom` data list. This parameter is
passed through from \code{gen_hw_set} and includes user specifications
for, if specified as other than the default, an alternative function to
use to identify heatwaves, alternative upper and lower year boundaries
for the projection period of the heatwave datasets being generated,
alternative upper and lower year boundaries for the reference period
to be used when measuring all heatwave characteristics related to
relative, rather than absolute, temperature, and the percentile
threshold to use when defining heatwaves.}

\item{reference}{FALSE, if the user has not specified custom reference
boundaries through the \code{referenceBoundaries} or
\code{dataBoundaries} arguments (depending on whether \code{experiment}
is "historical" or "rcp85") in
\code{gen_hw_set}, otherwise a vector with four elements
(earliest historical year, latest historical year,
earliest future projection year, latest future projection year)
giving the user-specified custom date boundaries for the
heatwave projections.}

\item{model}{List with parsed information about the directory structure for
a specific ensemble member from the user-specified projections directory.
This list is a subset of the list generated by
\code{acquireDirectoryStructure}.}
}
\value{
Returns a list with...
}
\description{
This function extract the desired climate data out of a single ensemble
   member of the climate model's projections either from a historical or
   future projection.
}
\details{
To do this, the function (1) reads in the latitude
   and longitude data from the grid location file for the ensemble member;
   (2) reads in the climate projection data for the ensemble member; (3)
   reads in the time file for the ensemble member; (4) finds the closest
   grid point in the climate model grid for each community in the
   \code{citycsv} file provided by the user; (5) ...
}
\note{
This function calls another function that uses the Pythagorean
   theorem to calculate the distance between each community given in the
   \code{citycsv} file and the nearest grid point for the climate model
   for the specified ensemble member.
}

