% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processModel}
\alias{processModel}
\title{Process valid models}
\usage{
processModel(model, global, custom, accumulators)
}
\arguments{
\item{model}{List with parsed information about the directory structure for
a specific climate model from the user-specified projections directory.
This list is a subset of the list generated by
\code{acquireDirectoryStructure}.}

\item{global}{An object name for the 'global' variable list. This
parameter is passed through from \code{\link{gen_hw_set}} and includes user
specifications for the path to the output directory, the path to the
input climate projections, the dataframe with community locations, the
format for the filenames of the coordinate, tas, and time files in the
ensemble-specific subdirectories of the projection directory input by
the user, and whether the user prefers to run the function using the R
or C++ version of the function to identify heatwaves.}

\item{custom}{An object name for the 'custom` data list. This parameter is
passed through from \code{\link{gen_hw_set}} and includes user specifications
for, if specified as other than the default, an alternative function to
use to identify heatwaves, alternative upper and lower year boundaries
for the projection period of the heatwave datasets being generated,
alternative upper and lower year boundaries for the reference period
to be used when measuring all heatwave characteristics related to
relative, rather than absolute, temperature, and the percentile
threshold to use when defining heatwaves.}

\item{accumulators}{An object name for the closure generated by
\code{\link{createAccumulators}}.}
}
\value{
A list object with the following elements: \code{locations}:
   A vector with the indices of the column of \code{tas} giving the
   climate projections for the climate grid point closest to each
   of the user-specified communities; \code{bounds}: indices linking
   the selected time bounds to the corresponding rows in the climate
   projection datafile, \code{tas} (first two values in the vector) as
   well as the total number of days within the range (third value);
   \code{series}: The climate projection data for the selected
   communities and time range. Each row corresponds to a different date;
   each column corresponds to a different community; \code{times}:
   A dataframe with the dates corresponding to each row of \code{series};
   \code{dates}: A vector with the dates, in Date format, corresponding
   to each row in \code{series}; \code{reference}: ... .
}
\description{
This function takes any models that are valid, identifies and characterizes
heatwaves within projections for each of its ensemble members, writes
those dataframes of heatwave projections out to the user-specified output
directory, and stores information about the models that have been written
to the output directory using the closure, \code{accumulators}, created by
\code{\link{createAccumulators}}.
}

