% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processModel}
\alias{processModel}
\title{Process valid models}
\usage{
processModel(model, global, custom, accumulators)
}
\arguments{
\item{model}{Character string of the climate model (e.g., "bcc1").}

\item{global}{List of global parameters generated by listGlobal.}

\item{custom}{List of custom parameters generated by listCustom.}

\item{accumulators}{Closures ... generated by createAccumulators}
}
\value{
[What does this function return?]
}
\description{
Process models that are valid and store information about the models
that have been written.
}

