% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processModel}
\alias{processModel}
\title{Process valid models}
\usage{
processModel(model, global, custom, accumulators)
}
\arguments{
\item{model}{Character string of the climate model (e.g., "bcc1").}

\item{global}{The 'global' variable list.}

\item{custom}{The user-specified custom data list.}

\item{accumulators}{Closures ... generated by createAccumulators}
}
\value{
A list of rcp 8.5 ensembles for the given model
}
\description{
Process models that are valid and store information about the models
that have been written.
}

