% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processModel}
\alias{processModel}
\title{Process valid models}
\usage{
processModel(model, global, custom, accumulators)
}
\arguments{
\item{model}{List with parsed information about the directory structure for
a specific climate model from the user-specified projections directory.
This list is a subset of the list generated by
\code{acquireDirectoryStructure}.}

\item{global}{An list object created by \code{\link{gen_hw_set}} that
includes user specifications (e.g., the path to the output directory, the
path to the input climate projections, the dataframe with community
locations).}

\item{custom}{An list object created by \code{\link{gen_hw_set}} that
includes user specifications (e.g., the name of the R function to
use to identify heatwaves, alternative upper and lower year boundaries
for the data used to determine threshold temperatures for the heatwave
definition, alternative upper and lower year boundaries
for the projection period of the heatwave datasets being generated).}

\item{accumulators}{An object name for the closure generated by
\code{\link{createAccumulators}}.}
}
\value{
A list object with processed projections, as well as various other
   elements needed for further processing to identify heat waves.
}
\description{
This function takes any models that are valid, identifies and characterizes
heatwaves within projections for each of its ensemble members, writes
those dataframes of heatwave projections out to the user-specified output
directory, and stores information about the models and grid locations.
}

