% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processModel}
\alias{processModel}
\title{Process valid models}
\usage{
processModel(model, global, custom, accumulators)
}
\arguments{
\item{model}{Character string of climate model name (e.g., "bcc1"). This
name is typically generated for use in this function from the subdirectory
names for the climate model within the directory of projection data
specified by the user in \code{gen_hw_set}.}

\item{global}{An object name for the 'global' variable list. This
parameter is passed through from \code{gen_hw_set} and includes user
specifications for the path to the output directory, the path to the
input climate projections, the dataframe with community locations, the
format for the filenames of the coordinate, tas, and time files in the
ensemble-specific subdirectories of the projection directory input by
the user, and whether the user prefers to run the function using the R
or C++ version of the function to identify heatwaves.}

\item{custom}{An object name for the 'custom` data list. This parameter is
passed through from \code{gen_hw_set} and includes user specifications
for, if specified as other than the default, an alternative function to
use to identify heatwaves, alternative upper and lower year boundaries
for the projection period of the heatwave datasets being generated,
and alternative upper and lower year boundaries for the reference period
to be used when measuring all heatwave characteristics related to
relative, rather than absolute, temperature.}

\item{accumulators}{An object name for the closure generated by
\code{createAccumulators}.}
}
\value{
A list of rcp 8.5 ensembles for the given model
}
\description{
This function takes any models that are valid, identifies and characterizes
   heatwaves within projections for each of its ensemble members, writes
   those dataframes of heatwave projections out to the user-specified output
   directory, and stores information about the models that have been written
   to the output directory using the closure, \code{accumulators}, created by
   \code{createAccumulators}.
}

