% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processRCP}
\alias{processRCP}
\title{Create heatwave dataframe for climate projection}
\usage{
processRCP(ensemble, modelName, ensembleWriter, thresholds, global, custom,
  accumulators, reference = FALSE)
}
\arguments{
\item{global}{The 'global' variable list.}

\item{custom}{The user-specified custom data list.}

\item{accumulators}{Closures ... generated by createAccumulators}

\item{reference}{[What is this parameter?]}
}
\value{
Write every heatwave dataframe to a .csv and return the ensemble
   used as the reference.
}
\description{
For each reference ensemble, create the heatwave dataframe and
write it to file. Also store the locations  vector for each ensemble.
}

