% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processThresholds}
\alias{processThresholds}
\title{Calculate threshold temperatures}
\usage{
processThresholds(model, global, custom)
}
\arguments{
\item{model}{List with parsed information about the directory structure for
a specific climate model from the user-specified projections directory.
This list is a subset of the list generated by
\code{acquireDirectoryStructure}.}

\item{global}{An list object created by \code{\link{gen_hw_set}} that
includes user specifications (e.g., the path to the output directory, the
path to the input climate projections, the dataframe with community
locations).}

\item{custom}{An list object created by \code{\link{gen_hw_set}} that
includes user specifications (e.g., the name of the R function to
use to identify heatwaves, alternative upper and lower year boundaries
for the data used to determine threshold temperatures for the heatwave
definition, alternative upper and lower year boundaries
for the projection period of the heatwave datasets being generated).}
}
\value{
A list with two elements: (1) a vector with one element for
   each community included in the user-specified community location
   file, with each element value giving the threshold temperature for the
   heatwave definition for a community, in the same order that communities
   are listed in the user-specificed \code{citycsv} file and (2) a
   \code{processModel} object, which is FALSE if the user did not
   specify custom year boundaries with \code{dataBoundaries} in
   \code{\link{gen_hw_set}} and otherwise is ... .
   This function also writes threshold temperatures to files within
   the user-specified output directory for future reference.
}
\description{
This function calculates the threshold temperatures required to identify
heatwaves in the climate projection data using the ensemble member specified
by \code{threshold_ensemble} in \code{\link{gen_hw_set}}. This threshold is
used in later functions to identify heatwaves.
}

