% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readLatLong}
\alias{readLatLong}
\title{Read latitude and longitude data}
\usage{
readLatLong(ensemble, global)
}
\arguments{
\item{ensemble}{Character vector that includes
the file paths to (1) the latitude and longitude file; (2) the
climate projection file; and (3) the projection dates file
for the selected climate model.}

\item{global}{An object name for the \code{global} object. This
object is created by \code{\link{gen_hw_set}} and includes user
specifications for the path to the output directory, the path to the
input climate projections, the dataframe with community locations, and the
format for the filenames of the coordinate, tas, and time files in the
ensemble-specific subdirectories of the projection directory input by
the user.}
}
\value{
A dataframe of the latitude and longitude data of the ensemble
   with columns for the latitude (first column) and longitude (second
   column) of each grid point in the climate model. This dataframe
   uses a format where the location of New York City, for example, is
   (40.7127, 285.9941).
}
\description{
This function reads data on the longitudes and latitudes of a climate
model into R from a locally-stored comma-separated file located within
the directory specified by the \code{dataFolder} argument of
\code{\link{gen_hw_set}}.
}

