% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readLatLong}
\alias{readLatLong}
\title{Read latitude and longitude data}
\usage{
readLatLong(ensemble, global)
}
\arguments{
\item{ensemble}{Character vector that includes
the file paths to (1) the latitude and longitude file; (2) the
climate projection file; and (3) the projection dates file
for the selected climate model.}

\item{global}{An list object created by \code{\link{gen_hw_set}} that
includes user specifications (e.g., the path to the output directory, the
path to the input climate projections, the dataframe with city
locations).}
}
\value{
A dataframe of the latitude and longitude data of the ensemble
   with columns for the latitude (first column) and longitude (second
   column) of each grid point in the climate model.
}
\description{
This function reads in data on the longitudes and latitudes of a climate
model's grid pooints from a locally-stored comma-separated file located within
the directory specified by the \code{dataFolder} argument of
\code{\link{gen_hw_set}}.
}

