% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readtas}
\alias{readtas}
\title{Read the time series data for a given ensemble}
\usage{
readtas(ensemble)
}
\arguments{
\item{ensemble}{Ensemble component of the directory structure that contains the path
to the time series data to be read}
}
\value{
A dataframe of the time series data of the ensemble
}
\description{
Read the time series data for a given ensemble
}

