% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readTimes}
\alias{readTimes}
\title{Read projection dates data}
\usage{
readTimes(ensemble, global)
}
\arguments{
\item{ensemble}{Character vector that includes
the file paths to (1) the latitude and longitude file; (2) the
climate projection file; and (3) the projection dates file
for the selected climate model.}

\item{global}{An list object created by \code{\link{gen_hw_set}} that
includes user specifications (e.g., the path to the output directory, the
path to the input climate projections, the dataframe with city
locations).}
}
\value{
A dataframe of dates corresponding to climate projection data,
   where each row gives a date, split into columns of day in the
   dataframe, four-digit year, numeric month, and numeric day.
}
\description{
This function reads the dates corresponding to climate projection data for
a climate model into R from a locally-stored comma-separated file located
within the directory specified by the \code{dataFolder} argument of
\code{\link{gen_hw_set}}.
}

