% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{storeHeatwaveEntry}
\alias{storeHeatwaveEntry}
\title{Helper function for IDHeatwavescpp that addes entries to the data structures that track
heatwave information}
\usage{
storeHeatwaveEntry(index, hwSize, hwCounter, hw, hwNumber)
}
\arguments{
\item{index}{An index into the various}

\item{hwSize}{Size of the heatwave to be added.}

\item{hwCounter}{Current number of heatwaves.}

\item{hw}{A reference of the vector that contains the heatwaves}

\item{hwNumber}{A reference of the vector that contains the heatwave numbers.}
}
\description{
Helper function for IDHeatwavescpp that addes entries to the data structures that track
heatwave information
}

