% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{storeHeatwaveEntry}
\alias{storeHeatwaveEntry}
\title{Helper function that adds entries}
\usage{
storeHeatwaveEntry(index, hwSize, hwCounter, hw, hwNumber)
}
\arguments{
\item{index}{A running index.}

\item{hwSize}{Size of the heat wave to be added.}

\item{hwCounter}{Current number of heat waves.}

\item{hw}{A reference of the vector that contains the heat waves}

\item{hwNumber}{A reference of the vector that contains the heat wave
numbers.}
}
\description{
This is a helper function for IDHeatwavesCPP that adds entries
to the data structures that track heatwave information
}

