% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{writeAccumulator}
\alias{writeAccumulator}
\title{Output the variable that accumulates information on the models and ensemble counts}
\usage{
writeAccumulator(modelInfoAccumulator, global)
}
\arguments{
\item{modelInfoAccumulator}{The dataframe that accumulates the number of ensembles for each model.}

\item{global}{The global data list}
}
\description{
Output the variable that accumulates information on the models and ensemble counts
}

