% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{writeAccumulator}
\alias{writeAccumulator}
\title{Write out model information}
\usage{
writeAccumulator(modelInfoAccumulator, global)
}
\arguments{
\item{modelInfoAccumulator}{The dataframe that accumulates the number of
ensembles for each model.}

\item{global}{An object name for the 'global' variable list. This
parameter is passed through from \code{\link{gen_hw_set}} and includes user
specifications for the path to the output directory, the path to the
input climate projections, the dataframe with community locations, and the
format for the filenames of the coordinate, tas, and time files in the
ensemble-specific subdirectories of the projection directory input by
the user.}
}
\description{
Output the variable that accumulates information on the models and
ensemble counts
}

