% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{writeAccumulators}
\alias{writeAccumulators}
\title{Write model information to file}
\usage{
writeAccumulators(modelInfoAccumulator, locationList, global)
}
\arguments{
\item{modelInfoAccumulator}{The dataframe that accumulates the number of
ensembles for each model.}

\item{global}{An object name for the \code{global} object. This
object is created by \code{\link{gen_hw_set}} and includes user
specifications for the path to the output directory, the path to the
input climate projections, the dataframe with community locations, and the
format for the filenames of the coordinate, tas, and time files in the
ensemble-specific subdirectories of the projection directory input by
the user.}
}
\value{
...
}
\description{
This function writes out a dataframe that accumulates information on all
climate models included in the analysis, including the number of
historical (1980--2004) and future (2006--2099) ensemble members.
}

